### 2020-09-17 #2679 @whedon
[PRE REVIEW]: Thermosteam: BioSTEAM's Premier Thermodynamic Engine
**Submitting author:** @yoelcortes (<a href="http://orcid.org/0000-0003-1742-5059">Yoel Cortes-Pena</a>)
**Repository:** <a href="https://github.com/BioSTEAMDevelopmentGroup/thermosteam" target ="_blank">https://github.com/BioSTEAMDevelopmentGroup/thermosteam</a>
**Version:** 0.20.26
