### 2020-09-17 #2679 @kyleniemeyer
Hey @openjournals/dev, I can see the paper in the `docs/paper` directory in the repo (https://github.com/BioSTEAMDevelopmentGroup/thermosteam/tree/master/docs/paper), any idea why whedon can't find it?
