### 2020-09-17 #2679 @kyleniemeyer
Hi @yoelcortes, thanks for your submission to JOSS! We'll use this issue to identify reviewers and resolve any technical issues (like our friendly bot not being able to find the paper).

I will be serving as editor for this; I have a few potential reviewers in mind, but any suggestions you have are also welcome.
