### 2020-09-17 #2679 @decaluwe
Happy to!

Steven

Sent from my iPhone

> On Sep 16, 2020, at 8:20 PM, Kyle Niemeyer <notifications@github.com> wrote:
>
> ﻿
> Hey @bryanwweber @decaluwe would one of you be interested in reviewing this for JOSS?
>
