### 2020-09-21 6eb90c2
minor changes for robust testing
