+++ .travis.yml
@@ -27,6 +27,6 @@ install:
   - python setup.py -q install
   - rm -rf thermosteam
 script:
-  - py.test . -v --cov-report html --cov=thermosteam --cov-report term-missing -m "not online and not thermo and not numba"
+  - py.test . -v --cov-report html --cov=thermosteam --cov-report term-missing -m "not online and not slow"
 after_success:
   - coveralls
