### 2020-10-05 1fa9091
fix bug with try_out method
