### 2020-10-05 1f4b2ed
make sure chemicals can be created without errors
