+++ tests/test_chemical.py
@@ -12,11 +12,9 @@ from numpy.testing import assert_allclose
 
 def test_chemical():
     import thermosteam as tmo
-    pubchem_db = tmo.chemicals.identifiers.pubchem_db
-    pubchem_db.load(pubchem_db.unloaded_files.pop())
-    for i in pubchem_db.CAS_index.values():
-        # Simply make sure they can be created without errors
-        tmo.Chemical(i.CASs)
+    CAS = ['12385-13-6', '7440-59-7', '7439-93-2', '7440-41-7', '7440-42-8', '7440-44-0', '17778-88-0', '17778-80-2', '14762-94-8', '7440-01-9', '7440-23-5', '7439-95-4', '7429-90-5', '7440-21-3', '7723-14-0', '7704-34-9', '22537-15-1', '7440-37-1', '7440-09-7', '7440-70-2', '7440-20-2', '7440-32-6', '7440-62-2', '7440-47-3', '7439-96-5', '7439-89-6', '7440-48-4', '7440-02-0', '7440-50-8', '7440-66-6', '7440-55-3', '7440-56-4', '7440-38-2', '7782-49-2', '10097-32-2', '7439-90-9', '7440-17-7', '7440-24-6', '7440-65-5', '7440-67-7', '7440-03-1', '7439-98-7', '7440-26-8', '7440-18-8', '7440-16-6', '7440-05-3', '7440-22-4', '7440-43-9', '7440-74-6', '7440-31-5', '7440-36-0', '13494-80-9', '7553-56-2', '7440-63-3', '7440-46-2', '7440-39-3', '7439-91-0', '7440-45-1', '7440-10-0', '7440-00-8', '7440-12-2', '7440-19-9', '7440-53-1', '7440-54-2', '7440-27-9', '7429-91-6', '7440-60-0', '7440-52-0', '7440-30-4', '7440-64-4', '7439-94-3', '7440-58-6', '7440-25-7', '7440-33-7', '7440-15-5', '7440-04-2', '7439-88-5', '7440-06-4', '7440-57-5', '7439-97-6', '7440-28-0', '7439-92-1', '7440-69-9', '7440-08-6', '7440-68-8', '10043-92-2', '7440-73-5', '7440-14-4', '7440-34-8', '7440-29-1', '7440-13-3', '7440-61-1', '7439-99-8', '7440-07-5', '7440-35-9', '7440-51-9', '7440-40-6', '7440-71-3', '7429-92-7', '7440-72-4', '7440-11-1', '10028-14-5', '22537-19-5', '53850-36-5', '53850-35-4', '54038-81-2', '54037-14-8', '54037-57-9', '54038-01-6', '54083-77-1', '54386-24-2', '54084-26-3', '54084-70-7', '54085-16-4', '54085-64-2', '54100-71-9', '54101-14-3', '54144-19-3', '78-96-6', '97-00-7', '107-06-2', '120-82-1', '107-20-0', '107-07-3', '100-54-9', '541-50-4', '123-08-0', '107-21-1', '64-19-7', '75-07-0', '60-35-5', '513-86-0', '67-64-1', '124-04-9', '7664-41-7', '100-52-7', '71-43-2', '65-85-0', '100-51-6', '10035-10-6', '123-72-8', '513-85-9', '71-36-3', '107-92-6', '462-94-2', '124-38-9', '630-08-0', '64-18-6', '120-80-9', '74-82-8', '79-11-8', '77-92-9', '7647-01-0', '95-48-7', '69-72-7', '108-39-4', '111-65-9', '124-07-2', '111-16-0', '79-46-9', '7783-06-4', 
[…]
