### 2020-10-20 de2487e
minor code style change
