+++ thermosteam/__init__.py
@@ -49,4 +49,4 @@ pd.set_option('display.max_columns', 10)
 pd.set_option('max_colwidth', 35)
 del np, pd
 
-__version__ = "0.20.28"
+__version__ = "0.20.29"
