### 2020-10-20 01814ba
add new copy method
