### 2020-10-20 0b9f6a8
make adjust_moisture_content more robust
