### 2020-10-25 #2679 @kyleniemeyer
Hi @fwitte, would you be interested in reviewing this submission for JOSS? Thank you!
