### 2020-10-29 b90ec65
implement absolute entropy (not just relative)
