+++ thermosteam/_chemical.py
@@ -29,10 +29,15 @@ from chemicals.triple import (Tt as triple_point_temperature,
 from chemicals.combustion import combustion_data, combustion_stoichiometry
 from chemicals.volume import volume_handle
 from chemicals.heat_capacity import heat_capacity_handle
-from chemicals.reaction import Hf as heat_of_formation
-from chemicals.elements import (similarity_variable as compute_similarity_variable, 
-                                simple_formula_parser as get_atoms,
-                                molecular_weight as compute_molecular_weight)
+from chemicals.reaction import (
+    Hf as heat_of_formation,
+    S0 as absolute_entropy_of_formation
+)
+from chemicals.elements import (
+    similarity_variable as compute_similarity_variable, 
+    simple_formula_parser as get_atoms,
+    molecular_weight as compute_molecular_weight
+)
 from chemicals.viscosity import viscosity_handle
 from chemicals.thermal_conductivity import thermal_conductivity_handle
 from chemicals.permittivity import permitivity_handle
@@ -171,11 +176,11 @@ _model_and_phase_properties = ('Psat', 'Hvap', 'sigma', 'epsilon',
 _handles = _model_and_phase_handles + _energy_handles
 
 _data = ('_MW', '_Tm', '_Tb', '_Tt', '_Tc', '_Pt', '_Pc', '_Vc',
-         '_Hf', '_LHV', '_HHV', '_Hfus', '_omega', '_dipole',
+         '_Hf', '_S0', '_LHV', '_HHV', '_Hfus', '_Sfus', '_omega', '_dipole',
          '_similarity_variable', '_iscyclic_aliphatic', '_combustion')
 
 _functor_data = {'MW', 'Tm', 'Tb', 'Tt', 'Pt', 'Tc', 'Pc', 'Vc',
-                 'Hfus', 'omega', 'dipole',
+                 'Hfus', 'Sfus', 'omega', 'dipole',
                  'similarity_variable', 'iscyclic_aliphatic'}
 
 _checked_properties = ('phase_ref', 'eos', 'eos_1atm',
@@ -284,9 +289,11 @@ class Chemical:
              Pc: 2.2048e+07 Pa
              Vc: 5.6e-05 m^3/mol
              Hf: -2.8582e+05 J/mol
+             S0: 70 J/K/mol
              LHV: 44011 J/mol
              HHV: 0 J/mol
              Hfus: 6010 J/mol
+             Sfus: None
              omega: 0.344
              dipole: 1.85 Debye
              similarity_variable: 0.16653
@@ -521,8 +528,6 @@ class Chemical:
     P_ref = 101325.
     #: [float] Reference enthalpy in J/mol
     H_ref = 0.
-    #: [float] Reference entropy in J/mol
-    S_ref = 0.
     #: dict[str, Chemical] Cached chemicals
     chemical_cache = {}
     #: [bool] Wheather or not to search cache by default
@@ -638,9 +643,11 @@ class Chemical:
                  Pc: None
                  Vc: None
                  Hf: None
+                 S0: None
                  LHV: None
                  HHV: None
                  Hfus: None
+                 Sfus: None
                  omega: None
                  dipole: None
                  similarity_variable: None
@@ -711,9 +718,11 @@ class Chemical:
                  Pc: 4.82e+06 Pa
                  Vc: 0.000414 m^3/mol
                  Hf: -1.2711e+06 J/mol
[…]
