### 2020-10-29 16b94c5
add units of measure for absolute entropy
