### 2020-10-29 fbea1c2
add chemical_splits function to find splits
