### 2020-10-29 7043ac2
update tests for absolute entropy values
