### 2020-10-29 1b6ed3e
make constants module public
