+++ thermosteam/chemicals/reaction.py
@@ -19,7 +19,7 @@ from chemicals import reaction
 import pandas as pd
 
 reaction.__all__.extend([
-    'Hf', 'Hf_at_phase'
+    'Hf', 'S0', 'free_energy_at_phase'
 ])
 
 folder = os.path.join(os.path.dirname(__file__), 'Reaction')
@@ -65,7 +65,7 @@ def Hf(CASRN, phase=None, Hvap=None, Hfus=None):
         Hf = Hf_biochemicals.at[CASRN, 'Hf']
         phase_found = Hf_biochemicals.at[CASRN, 'phase']
         if phase and phase_found != phase: 
-            Hf = Hf_at_phase(Hf, phase_found, phase, Hvap, Hfus)
+            Hf = free_energy_at_phase(Hf, phase_found, phase, Hvap, Hfus)
         return Hf
     if not phase:
         for f in (reaction.Hfg, reaction.Hfl, reaction.Hfs):
@@ -81,51 +81,109 @@ def Hf(CASRN, phase=None, Hvap=None, Hfus=None):
         raise ValueError("phase must be either 's', 'l', or 'g'; not '%s'" %phase)
     if Hf is None:
         if phase != 'g':
-            Hf = Hf_at_phase(reaction.Hfg(CASRN), 'g', phase, Hvap, Hfus)
+            Hf = free_energy_at_phase(reaction.Hfg(CASRN), 'g', phase, Hvap, Hfus)
         elif phase != 'l':
-            Hf = Hf_at_phase(reaction.Hfl(CASRN), 'l', phase, Hvap, Hfus)
+            Hf = free_energy_at_phase(reaction.Hfl(CASRN), 'l', phase, Hvap, Hfus)
         elif phase != 's':
-            Hf = Hf_at_phase(reaction.Hfs(CASRN), 's', phase, Hvap, Hfus)
+            Hf = free_energy_at_phase(reaction.Hfs(CASRN), 's', phase, Hvap, Hfus)
     return Hf
 reaction.Hf = Hf
 
-def Hf_at_phase(Hf, phase_ref, phase, Hvap, Hfus):
+def S0(CASRN, phase=None, Svap=None, Sfus=None):
+    r'''
+    Return a chemical's absolute entropy of formation at given phase. If no phase given,
+    return the standard-state entropy of formation. Return None if the data is 
+    not available. 
+    
+    Parameters
+    ----------
+    CASRN : string
+        CASRN [-].
+    
+    Returns
+    -------
+    S0 : float
+        Absolute entropy of formation [J/mol].
+    
+    Other Parameters
+    ----------------
+    phase : 's', 'l', or 'g', optional
+        Phase.
+    Svap=None : float, optional
+        Entropy of vaporization [J/mol].
+    Sfus=None : float, optional
+        Entropy of fusion [J/mol].
+    
+    Notes
+    -----
+    This is a wrapper around :func:`~chemicals.reaction.S0s`,
+    :func:`~chemicals.reaction.S0l`, and :func:`~chemicals.reaction.S0g`.
+    
+    Examples
+    --------
+    >>> S0(CASRN='7732-18-5')
+    188.8
+    
+    '''
+    if not phase:
+        for f in (reaction.S0g, reaction.S0l, reaction.S0s):
+            S0 = f(CASRN)
+            if S0: break 
+    elif phase == 'g':
+        S0 = reaction.S0g(CASRN)
+    elif phase == 'l':
+        S0 = reaction.S0l(CASRN)
+    elif phase == 's':
+        S0 = reaction.S0s(CASRN)
+    else:
+        raise ValueError("phase must be either 's', 'l', or 'g'; not '%s'" %phase)
[…]
