### 2020-10-30 6540c52
make phase management more robust
