### 2020-10-31 35aaa81
fix bug to return Sfus (not Hfust)
