### 2020-10-31 a232211
make stream links more robust
