### 2020-11-01 0437f19
add work around for copy_like when multiple phases are present
