### 2020-11-01 e79e908
make sure phase property returns a string
