### 2020-11-02 #2814 @decaluwe
Hi @kyleniemeyer - I made it all the way to step 2 🤓

The link https://github.com/openjournals/joss-reviews/invitations takes me to the following message:

>Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account.
