### 2020-11-03 BioSTEAMDevelopmentGroup/thermosteam#23 @fwitte
Very large download size of the git repo
Cloning the repo took quite a while as some files in the git history seem to be very large (total repo size > 1.5 GB, which seems excessive to me).
This is not part of the JOSS paper review, but I think, it could be a good idea, to see what one can do about this. Unfortunately, I cannot help you here as I have no experience on this topic, other than that history manipulation usually is considered bad. Basically, just wanted to let you know.
Best regards
Francesco
