### 2020-11-03 da26192
decrease memory usage for readthedocs by making scipy optional, and mocking import
