+++ setup.py
@@ -11,17 +11,23 @@ setup(
     name='thermosteam',
     packages=['thermosteam'],
     license='MIT',
-    version='0.21.1',
+    version='0.21.2',
     description="BioSTEAM's Premier Thermodynamic Engine",
     long_description=open('README.rst').read(),
     author='Yoel Cortes-Pena',
-    install_requires=['pint>=0.9', 'chemicals>=0.1.4',
-                      'scipy>=1.3.1', 'IPython>=7.9.0', 
-                      'colorpalette>=0.3.1', 'biosteam>=2.21.1',
-                      'pandas>=0.25.2', 'matplotlib>=3.1.1',
-                      'numpy>=1.18.1', 'xlrd==1.2.0',
-                      'openpyxl>=3.0.0', 'free_properties>=0.2.5',
-                      'flexsolve==0.3.32', 'pyglet',
+    install_requires=['pint>=0.9', 
+                      'chemicals>=0.1.4',
+                      'IPython>=7.9.0', 
+                      'colorpalette>=0.3.1', 
+                      'biosteam>=2.21.1',
+                      'pandas>=0.25.2', 
+                      'matplotlib>=3.1.1',
+                      'numpy>=1.18.1', 
+                      'xlrd==1.2.0',
+                      'openpyxl>=3.0.0', 
+                      'free_properties>=0.2.5',
+                      'flexsolve==0.3.32', 
+                      'pyglet',
                       'numba>=0.50.0'],
     package_data={
         'thermosteam': [
