+++ requirements.txt
@@ -1,6 +1,5 @@
 pint>=0.9
 chemicals>=0.1.4
-scipy>=1.3.1
 ipython>=7.12.0
 colorpalette>=0.3.1
 pandas>=0.25.2
