+++ setup.py
@@ -11,7 +11,7 @@ setup(
     name='thermosteam',
     packages=['thermosteam'],
     license='MIT',
-    version='0.21.2',
+    version='0.21.3',
     description="BioSTEAM's Premier Thermodynamic Engine",
     long_description=open('README.rst').read(),
     author='Yoel Cortes-Pena',
@@ -19,7 +19,7 @@ setup(
                       'chemicals>=0.1.4',
                       'IPython>=7.9.0', 
                       'colorpalette>=0.3.1', 
-                      'biosteam>=2.21.1',
+                      'biosteam>=2.21.4',
                       'pandas>=0.25.2', 
                       'matplotlib>=3.1.1',
                       'numpy>=1.18.1', 
