### 2020-11-03 23b439a
update chemicals requirements for readthedocs build
