### 2020-11-03 fd870bb
add scipy back to requirements for numba
