+++ setup.py
@@ -11,15 +11,16 @@ setup(
     name='thermosteam',
     packages=['thermosteam'],
     license='MIT',
-    version='0.21.4',
+    version='0.21.5',
     description="BioSTEAM's Premier Thermodynamic Engine",
     long_description=open('README.rst').read(),
     author='Yoel Cortes-Pena',
-    install_requires=['pint>=0.9', 
+    install_requires=['pint>=0.9',
+                      'scipy>=1.3.1', 
                       'chemicals>=0.1.5',
                       'IPython>=7.9.0', 
                       'colorpalette>=0.3.1', 
-                      'biosteam>=2.21.5',
+                      'biosteam>=2.21.6',
                       'pandas>=0.25.2', 
                       'matplotlib>=3.1.1',
                       'numpy>=1.18.1', 
