### 2020-11-03 91601da
fix energy balance bug with mixing and add tests
