+++ thermosteam/__init__.py
@@ -50,4 +50,4 @@ pd.set_option('display.max_columns', 10)
 pd.set_option('max_colwidth', 35)
 del np, pd
 
-__version__ = "0.21.5"
+__version__ = "0.21.6"
