### 2020-11-03 BioSTEAMDevelopmentGroup/thermosteam#22 @yoelcortes
Thanks for the bug report! It's been taken care of. You should get the following using your code:

```
MultiStream with enthalpy added by mix_from
MultiStream: s34_mixture
 phases: ('g', 'l'), T: 372.79 K, P: 100000 Pa
 composition: (g) Water  1
                  -----  0.497 kmol/hr
              (l) Water  1
                  -----  0.614 kmol/hr
Absolute deviation in enthalpy: -2.382876118645072e-09
MultiStream with Enthalpy summed individually
MultiStream: s34_mixture
 phases: ('g', 'l'), T: 372.79 K, P: 100000 Pa
 composition: (g) Water  1
                  -----  0.497 kmol/hr
              (l) Water  1
                  -----  0.614 kmol/hr
```

Edit: You can "pip install thermosteam==0.21.6" for convenience or use the github repo
