### 2020-11-03 BioSTEAMDevelopmentGroup/thermosteam#23 @yoelcortes
Hi Francesco, thanks for noting this. Yes! The repo is huge mostly because, historically, thermosteam incorporated all the data from [thermo](https://github.com/CalebBell/thermo) due to some problems with simply importing the data from [thermo](https://github.com/CalebBell/thermo).  Now all the data is imported from [chemicals](https://github.com/CalebBell/chemicals), a package that exposes just the databanks and functions for better code-reuse.

I am planning to remove the history before August 2020, but want to do so after the review (just in case reviewers would like to see the commits).

Edit: I'll keep this issue open until I remove old history.

Thanks!
