### 2020-11-03 29ecc2e
make mix_from more robust when equilibrium is present
