### 2020-11-07 345f568
fix LLE bugs, update examples, and minor changes
