### 2020-11-09 #2814 @decaluwe
I am currently having trouble installing the software.  Have tried multiple pathways using pip, to no avail.  Plan to try the GitHub source code, today...
