### 2020-11-09 #2814 @fwitte
I started into it. Will most likely complete the review this weekend or the following week. I am posting all software related issues at https://github.com/BioSTEAMDevelopmentGroup/thermosteam/ and will write a complete feedback here linking the respective issues.
