### 2020-11-09 #2814 @yoelcortes
>
>
> I am currently having trouble installing the software. Have tried multiple pathways using pip, to no avail. Plan to try the GitHub source code, today...

Sorry to hear that! It might be a bug with the Python's new pip installer. But in the case it is not, I uploaded the latest version to PyPI (pip install thermosteam==0.21.7). Hope this fixes the issue.

Thanks!
