### 2020-11-10 #2814 @decaluwe
2 quick questions:

- @kyleniemeyer if they don't list all dependencies, but the installation is handled by an automated package manager, am I fine to check this item off?

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

- @yoelcortes I see that there is a test suite in the source code--are there instructions for running it?

Thanks.
