### 2020-11-10 #2814 @yoelcortes
Hi @decaluwe, thanks for asking! If you have the github repository downloaded, go to your local thermosteam repository and run "pytest" in your cmd or terminal. Here are some instructions: https://biosteam.readthedocs.io/en/latest/CONTRIBUTING.html. Unfortunately I haven't added a note of this in Thermosteam's documentation yet.

I'd also like to note that continuous integration is built in on github, so that the package is installed in a virtual environment (on python 3.6, 3.7 and 3.8) and all the documentation examples and the test files are run with every push to the repository (the coverage is checked and updated as well). If you click on the red "X" mark, you can go to several links for details regarding tests of the last push:

![image](https://user-images.githubusercontent.com/29529649/98686318-06885d80-232e-11eb-96b6-ea966f436b55.png)

As for listing dependencies, would the "requirements.txt or setup.py" file count? Note that if you downloaded the repository, you can use "pip install ." to let pip install thermosteam and its dependencies for you as well.
