### 2020-11-10 #2814 @fwitte
Hi @decaluwe and @kyleniemeyer,
according to https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions I understood a list is not necessary with automatic installation of dependencies. Therefore I checked the box. Let me know if I did wrong.
@yoelcortes, maybe consider adding the 'extras_require' keyword to your setup.py, so developers will automatically have develop dependencies installed and do not need to install pytest, sphinx, ..., manually for testing/debugging on their respective virtual environment.
With
``` bash
    extras_require={'dev': ['pytest', 'sphinx', 'sphinx_rtd_theme', ]}
```
in setup py
``` bash
pip install -e .[dev]
```
will install the extras.

Best regards
Francesco
