### 2020-11-10 #2814 @kyleniemeyer
From my perspective, the dependency requirement is satisfied for Python packages through including appropriately in `setup.pu` and/or `requirements.txt` (for example), as @fwitte mentioned
