### 2020-11-11 #2814 @decaluwe
Okay, thanks @kyleniemeyer .  Since it was listed under the `documentation` section, I wasn't sure whether it needed to be explicitly listed, or if having it in `requirements.txt` was sufficient (since it is a PyPI-managed installation, it certainly seems sufficient to me).
