### 2020-11-16 #2814 @decaluwe
FYI:

1. @yoelcortes I get two test failures:
==================================== FAILURES================================
_____________________________________________ test_wheatstraw ______________________________________________

    @pytest.mark.slow
    def test_wheatstraw():
>       from biorefineries import wheatstraw as ws
E       ImportError: cannot import name 'wheatstraw' from 'biorefineries' (/Users/decaluwe/.local/lib/python3.7/site-packages/biorefineries/__init__.py)

../../../.local/lib/python3.7/site-packages/biorefineries/tests/test_biorefineries.py:74: ImportError
_____________________________________________ test_annimal_bedding ________________________________________

    @pytest.mark.slow
    def test_annimal_bedding():
>       from biorefineries import animal_bedding as ab
E       ImportError: cannot import name 'animal_bedding' from 'biorefineries' (/Users/decaluwe/.local/lib/python3.7/site-packages/biorefineries/__init__.py)

../../../.local/lib/python3.7/site-packages/biorefineries/tests/test_biorefineries.py:90: ImportError
============================ 2 failed, 153 passed in 12.11s =========================

2. @kyleniemeyer, is the automated testing check box require that the tests pass, or simply that the test suite exists?
