### 2020-11-16 #2814 @decaluwe
@yoelcortes Notes on the software paper:

1. Depending on one's definition of "a diverse, non-specialist audience," the intro could probably be written a little bit broader. For example, phrases like "unit operations" and even "chemical process" might be opaque for somebody in materials science or astronomy.  I'd suggest something like (i) end the first sentence right before the parenthetical, (ii) describe some common features of a "unit operation in a chemical process" (e.g. something along the lines of "fluid flows enter a reactor wherein chemical reactions occur, consuming or producing chemical species and thermal energy...") and then (iii) give some examples/categories of the specific types of properties needed for evaluating the design of a chemical process.  It might add a sentence or two, but should be clearer for a non-expert.

The distillation column example provides a good specific example, but something helping bridge the very general first sentence and this more specific example would be helpful.

2. State of the field: There is no documenting of or comparison to current capabilities in the field.  Would software like EES, CoolProp, RefProp, Cantera, or Chemkin be valid points of comparison? These packages might have similar functionality, but I do not know if they are "commonly used" in the field (which is what the review guidelines specifically ask), so I'd rely on your expertise, obviously.  What approaches are commonly used in your field?

Not related to the paper, but a general compliment: the API documentation is very well done, and is well populated with examples, which I appreciated.  Nice work!
