### 2020-11-17 0764b96
include extras_require for developer version
