### 2020-11-17 #2814 @yoelcortes
@decaluwe, I made some minor changes in the repository that prevents the pytest failures (thanks for reporting). Also, your comments on the software paper are good points. I'll wait until Francesco can submit his comments too before I submit my edits.

Glad you liked the documentation. I was quite inspired by (and grateful of) other open source projects with excellent documentation. Thanks!
