### 2020-11-17 BioSTEAMDevelopmentGroup/thermosteam#28 @fwitte
Correct_atomic_balance does not work on methane combustion equation
Hi Yoel,
I always like to try and fiddle with new code/software and wanted to create something I am more familiar with than biodiesel reactions :), a combustion reaction of methane with auto correcting the atomic balance. Unfortunately, I do not really know what I am doing wrong here, can you help me? I would expect, that the atomic balance can be solved here.

``` python
import thermosteam as tmo
from thermosteam import reaction as rxn

chemicals = tmo.Chemicals(['Water', 'Methane', 'CO2', 'O2'])
tmo.settings.set_thermo(chemicals)
combustion = rxn.Reaction('Methane + O2 -> Water + CO2',
                                   reactant='Methane', X=1,
                                   correct_atomic_balance=True)
```

Thanks and best regards
Francesco
