### 2020-11-17 BioSTEAMDevelopmentGroup/thermosteam#29 @fwitte
Heat of reaction does not account for state of water (liquid/gaseous) after reaction
Hi Yoel,
again looking at methane combustion. Feeding a stream of Methane and O2 into a reaction at a temperature of T = 100 °C (could be less, I wanted to make sure after reaction temperature is larger than saturation temperature at partial pressure of the water) the outlet temperature is the same and therefore water should be gaseous.

Is there a way to calculate the actual heat of reaction in respect to the before/after reaction state? Or is the heat of reaction not supposed to be calculated on this basis?

``` python
import thermosteam as tmo
from thermosteam import reaction as rxn

chemicals = tmo.Chemicals(['Water', 'Methane', 'CO2', 'O2', 'Glucose', 'Ethanol'])
tmo.settings.set_thermo(chemicals)
combustion = rxn.Reaction('Methane + 2O2 -> 2Water + CO2', reactant='Methane', X=1)

heat_of_combustion = combustion.dH / chemicals['Methane'].MW

feed = tmo.Stream(Methane=1, O2=2, T=373.15, phase='g')
print('BEFORE REACTION')
feed.show(N=100)

# React feed molar flow rate
combustion(feed)

print('AFTER REACTION')
feed.show(N=100)
print('This result is zero, if the state of the fluids is not considered in the heat of reaction:')
print(heat_of_combustion - combustion.dH / chemicals['Methane'].MW)
```

Thank you very much and best regards
Francesco
