### 2020-11-17 #2814 @fwitte
Hi @yoelcortes,
here are my general comments to the paper and your software (documentation).

## Paper

- The paper highlights the most important features of the software, but (as already written in the docs issue) I do miss the capability, that generic chemical reactions can be computed with your software. I think this is a very important feature of your software.
- A statement regarding the main advantages compared to software in the field (like @decaluwe pointed out) would be very good to emphasize the value of thermosteam, too.
- I am not a legal expert, is it necessary to include the disclaimer twice. Can it not be handled in a single sentence?
  Maybe something like "Any opinions, findings, and conclusions or recommendations expressed in this publication are those of the author and necessarily reflect the views of the funders and sponsors."?
- Other than that, good compact overview of the software.

## Software (Documentation)

There is not much to say here, very well documented with an extensive tutorial on how to use the package. I like it very much!

## Issues
Checked, if resolved.

- [ ] (optional) https://github.com/BioSTEAMDevelopmentGroup/thermosteam/issues/21
- [x] https://github.com/BioSTEAMDevelopmentGroup/thermosteam/issues/22
- [ ] (optional) https://github.com/BioSTEAMDevelopmentGroup/thermosteam/issues/23
- [ ] https://github.com/BioSTEAMDevelopmentGroup/thermosteam/issues/28
- [ ] https://github.com/BioSTEAMDevelopmentGroup/thermosteam/issues/29
- [ ] https://github.com/BioSTEAMDevelopmentGroup/thermosteam/issues/30
