### 2020-11-17 b593af8
add sympy to extras
