### 2020-11-17 41127dc
add methane combustion example to correct atomic balance method
