### 2020-11-17 BioSTEAMDevelopmentGroup/thermosteam#28 @yoelcortes
Thanks so much for testing this out. I enhanced the correct_atomic_balance method to be more flexible (check out the latest git version or "pip install thermosteam==0.21.8). Your test should work now:

```python
>>> import thermosteam as tmo
>>> from thermosteam import reaction as rxn
>>> chemicals = tmo.Chemicals(['Water', 'Methane', 'CO2', 'O2'])
>>> tmo.settings.set_thermo(chemicals)
>>> combustion = rxn.Reaction('Methane + O2 -> Water + CO2',
...                                reactant='Methane', X=1,
...                                correct_atomic_balance=True)
>>> combustion.show()
Reaction (by mol):
 stoichiometry                    reactant    X[%]
 Methane + 2 O2 -> 2 Water + CO2  Methane   100.00
```

I updated the documentation with more details too:
https://thermosteam.readthedocs.io/en/latest/reaction/Reaction.html#thermosteam.reaction.Reaction.correct_atomic_balance

Let me know if you have any further questions!
