### 2020-11-17 ba1608d
enhance documentation for correct_atomic_balance
