### 2020-11-17 ffc48c3
further enhance documentation for correct_atomic_balance
