### 2020-11-17 2f7f828
enhance adiabatic reaction documentation
