### 2020-11-17 BioSTEAMDevelopmentGroup/thermosteam#29 @yoelcortes
Hi Francesco,

I am not too sure if I understand your question correctly, but I think I can help clarify some details (that are not too apparent) and hopefully answer it in the process.

The [dH](https://thermosteam.readthedocs.io/en/latest/reaction/Reaction.html#thermosteam.reaction.Reaction.dH) property is the heat of reaction in J/mol-reactant. It takes into account the conversion of the reactant too, so dH at X=0.5 is half dH at X=1.

Calling the Reaction object with a Stream object assumes an isothermal reaction (as you already know). For an adiabatic reaction, you can use the [adiabatic_reaction](https://thermosteam.readthedocs.io/en/latest/reaction/Reaction.html#thermosteam.reaction.Reaction.adiabatic_reaction) method:

The following example shows that the flow rate of reactant multiplied by dH is equal to the total heat produced from the reaction:

```python
import thermosteam as tmo
from thermosteam import reaction as rxn

chemicals = tmo.Chemicals(['Water', 'Methane', 'CO2', 'O2', 'N2'])
tmo.settings.set_thermo(chemicals)
combustion = rxn.Reaction('Methane + 2O2 -> 2Water + CO2', reactant='Methane', X=1)
heat_of_combustion = combustion.dH

feed = tmo.Stream(Methane=1, O2=2, N2=100, T=373.15, phase='g') # Dilute with nitrogen to keep temperature change low
reactant_mol = feed.imol[combustion.reactant]
H0 = feed.H # Enthalpy not including formation enthalpies
Hnet0 = feed.Hnet # Enthalpy including formation enthalpies
print('BEFORE ADIABATIC REACTION')
feed.show(N=100)

# React feed molar flow rate
combustion.adiabatic_reaction(feed)

print('AFTER ADIABATIC REACTION')
feed.show(N=100)
Hf = feed.H
Hnetf = feed.Hnet
heat_gained = Hf - H0
dHnet = Hnetf - Hnet0
print(f"Error in energy conservation: {dHnet:.2f}")
print(f'Heat of combution {reactant_mol * heat_of_combustion:.2f} is equal to the heat gained by the system {heat_gained:.2f}')
```

Output:
```
BEFORE ADIABATIC REACTION
Stream: s21
 phase: 'g', T: 373.15 K, P: 101325 Pa
 flow (kmol/hr): Methane  1
                 O2       2
                 N2       100
AFTER ADIABATIC REACTION
Stream: s21
 phase: 'g', T: 631.91 K, P: 101325 Pa
 flow (kmol/hr): Water  2
                 CO2    1
                 N2     100
Error in energy conservation: 0.00
Heat of combution -890590.00 is equal to the heat gained by the system 890590.00
```

I hope I didn't complicate things more, please let me know if I haven't answered your question (or if you have anymore).

Thanks!
