### 2020-11-17 BioSTEAMDevelopmentGroup/thermosteam#29 @yoelcortes
Hi Francesco,

I just realized what you meant after reading the title of the issue! Sorry, there is no direct way of taking the state of chemicals. You could use a multistream to change the state of water after an adiabatic reaction and compute the enthalpy difference. If you are just working with combustion reactions, I would suggest you use the "LHV" and "HHV" properties of streams and chemicals:

```python
import thermosteam as tmo

chemicals = tmo.Chemicals(['Water', 'Methane', 'CO2', 'O2', 'N2'])
tmo.settings.set_thermo(chemicals)

feed = tmo.Stream(Methane=1, O2=2, N2=100, T=373.15, phase='g')
print(f'Lower heating value: {feed.LHV} [kJ/hr]')
print(f'Higher heating value: {feed.HHV} [kJ/hr]')
print(f'Lower heating value of methane: {chemicals.Methane.LHV} [J/mol]')
print(f'Higher heating value of methane: {chemicals.Methane.HHV} [J/mol]')
```

Output:
```
Lower heating value: -802567.008 [kJ/hr]
Higher heating value: -890590.0 [kJ/hr]
Lower heating value of methane: -802567.008 [J/mol]
Higher heating value of methane: -890590.0 [J/mol]
```

Hope this helps,
Thanks!
