### 2020-11-18 BioSTEAMDevelopmentGroup/thermosteam#29 @fwitte
Hi Yoel,
thanks for your detailed reply. I was talking about LHV/HHV :), I should have written that more clearly.

I think adding a sentence or two the corresponding parts of your documentation, where you state, that the heat of reaction is not considering the state of the reaction educts (the same "issue" happens when modeling water electrolysis with evaporated water as input) and reaction products.

Thank you, have a nice day!
Francesco
