### 2020-11-18 9b3317f
increase code-reuse and enhance documentation
