### 2020-11-19 af7a8e8
enhance documentation and fix problems addressed by JOSS reviewers
