+++ docs/conf.py
@@ -63,7 +63,7 @@ templates_path = ['_templates']
 # You can specify multiple suffix as a list of string:
 #
 # source_suffix = ['.rst', '.md']
-source_suffix = ['.txt', '.ipynb']
+source_suffix = ['.rst', '.txt', '.ipynb']
 
 # The master toctree document.
 master_doc = 'index'
