+++ thermosteam/__init__.py
@@ -31,13 +31,14 @@ from . import (indexer,
 from ._stream import Stream
 from ._multi_stream import MultiStream
 from .base import functor
+from .reaction import *
 from flexsolve import speed_up
 
-__all__ = ('Chemical', 'Chemicals', 'CompiledChemicals', 'Thermo', 'indexer',
-           'Stream', 'MultiStream', 'ThermalCondition', 'mixture', 'ThermoData',
-           'settings', 'functor', 'functors', 'chemicals', 'base', 'equilibrium',
-           'units_of_measure', 'exceptions', 'functional', 'reaction', 'constants',
-           'utils', 'separations', 'speed_up')
+__all__ = ('Chemical', 'Chemicals', 'CompiledChemicals', 'Thermo', 'Stream',
+           'MultiStream', 'ThermalCondition', 'mixture', 'ThermoData',
+           *reaction.__all__, 'indexer', 'settings', 'functor', 'functors', 
+           'chemicals', 'base', 'equilibrium', 'units_of_measure', 'exceptions',
+           'functional', 'reaction', 'constants', 'utils', 'separations', 'speed_up')
 
 # Set number of digits displayed
 import numpy as np
@@ -50,4 +51,4 @@ pd.set_option('display.max_columns', 10)
 pd.set_option('max_colwidth', 35)
 del np, pd
 
-__version__ = "0.21.10"
+__version__ = "0.22.1"
