### 2020-11-19 BioSTEAMDevelopmentGroup/thermosteam#30 @fwitte
You're welcome, thanks for updating! I will close this for now and reopen, if I spot something more.

Actually, going through your documentation, I really think the use of ipynb in readthedocs is very nice. I am considering changing parts of the documentation in my software packages to ipynb, too :). Just need to find the time or readjust priorities :).
