### 2020-11-20 83f9ee2
implement phase changes in Reaction objects, without documentation (yet)
