+++ thermosteam/reaction/_xparse.py
@@ -5,14 +5,39 @@
 # This module is under the UIUC open-source license. See 
 # github.com/BioSTEAMDevelopmentGroup/biosteam/blob/master/LICENSE.txt
 # for license details.
-
 import numpy as np
+from .._phase import valid_phases
 
-__all__ = ('get_stoichiometric_array',
+__all__ = ('get_phases',
+           'get_stoichiometric_array',
            'get_stoichiometric_string',
            'str2dct', 'dct2str', 'arr2dct')
 
-def get_stoichiometric_array(reaction, chemicals):
+def get_phases(reaction):
+    """Return all available phases in a reaction string."""
+    isa = isinstance
+    if isa(reaction, dict):
+        phases = []
+        for i, j in reaction.items():
+            if isa(j, str): return None
+            phase = j[0]
+            if phase not in valid_phases:
+                raise ValueError(f'invalid phase {repr(phase)} encountered while parsing reaction')
+            phases.append(phase)
+    elif isa(reaction, str):
+        phases = []
+        for i, x in enumerate(reaction):
+            try: phase = reaction[i+1]
+            except: break
+            if x == ',':
+                if phase not in valid_phases:
+                    raise ValueError(f'invalid phase {repr(phase)} encountered while parsing reaction')
+                phases.append(phase)
+        return tuple(phases)
+    else:
+        raise ValueError(f"reaction must be either a str or a dict, not a '{type(reaction).__name__}' object")
+
+def get_stoichiometric_array(reaction, phases, chemicals):
     """Return stoichiometric array given a string defining the reaction and chemicals."""
     if isinstance(reaction, dict):
         stoichiometry_dict = reaction
@@ -20,24 +45,25 @@ def get_stoichiometric_array(reaction, chemicals):
         stoichiometry_dict = str2dct(reaction)
     else:
         raise ValueError(f"reaction must be either a str or a dict, not a '{type(reaction).__name__}' object")
-    stoichiometric_array = dct2arr(stoichiometry_dict, chemicals)
+    stoichiometric_array = dct2arr(stoichiometry_dict, phases, chemicals)
     return stoichiometric_array
 
-def get_stoichiometric_string(reaction, chemicals):
+def get_stoichiometric_string(reaction, phases, chemicals):
     """Return a string defining the reaction given the stoichiometric array and chemicals."""
     if isinstance(reaction, dict):
         stoichiometric_dict = reaction
     elif isinstance(reaction, np.ndarray):
-        stoichiometric_dict = arr2dct(reaction, chemicals)
+        stoichiometric_dict = arr2dct(reaction, phases, chemicals)
     else:
         raise ValueError(f"reaction must be either a str or a dict, not a '{type(reaction).__name__}' object")
     return dct2str(stoichiometric_dict)
 
-def dct2arr(dct, chemicals):
-    arr = np.zeros(chemicals.size)
+def dct2arr(dct, phases, chemicals):
+    phase_index = {j:i for i,j in enumerate(phases)}
+    arr = np.zeros([len(phases), chemicals.size])
     chemical_index = chemicals._index
-    for ID, coefficient in dct.items():
-        arr[chemical_index[ID]] = coefficient
+    for ID, (phase, coefficient) in dct.items():
+        arr[phase_index[phase], chemical_index[ID]] = coefficient
     return arr 
 
 def split_coefficient(nID, sign):
@@ -54,12 +80,20 @@ def split_coefficient(nID, sign):
 def extract_coefficients(nIDs, dct, sign):
     for nID in nIDs:
         n, ID = split_coefficient(nID, sign)
+        if ID[-2] == ',': 
+            phase = ID[-1]
[…]
