+++ thermosteam/reaction/_reaction.py
@@ -11,14 +11,26 @@ import thermosteam as tmo
 import flexsolve as flx
 from chemicals import elements
 from warnings import warn
-from . import _parse as prs
+from . import (
+    _parse as prs,
+    _xparse as xprs,
+)
 from ..utils import chemicals_user
-from .._phase import NoPhase
-from ..indexer import ChemicalIndexer
+from .._phase import NoPhase, AnyPhase
+from ..indexer import ChemicalIndexer, MaterialIndexer
 from ..exceptions import InfeasibleRegion
 import numpy as np
 
 __all__ = ('Reaction', 'ParallelReaction', 'SeriesReaction')
+
+def get_stoichiometric_string(reaction):
+    if reaction.any_phase:
+        return prs.get_stoichiometric_string(reaction._stoichiometry, 
+                                             reaction._chemicals)
+    else:
+        return xprs.get_stoichiometric_string(reaction._stoichiometry,
+                                              reaction._phases,
+                                              reaction._chemicals)
 
 def react_stream_adiabatically(stream, reaction):
     if not isinstance(stream, tmo.Stream):
@@ -41,18 +53,19 @@ def set_reaction_basis(rxn, basis):
         rxn._rescale()
         rxn._basis = basis
 
-def as_material_array(material, basis, chemicals):
+def as_material_array(material, basis, phases, chemicals):
     isa = isinstance
     if isa(material, np.ndarray):
         return material
-    elif isa(material, tmo.MultiStream):
-        raise ValueError('reacting a stream with multiple phases is underdetermined')
     elif isa(material, tmo.Stream):
-        assert material.chemicals is chemicals, "reaction and stream chemicals do not match"
+        if material.phases != phases:
+            raise ValueError("reaction and stream phases do not match")
+        if material.chemicals is not chemicals:
+            raise ValueError("reaction and stream chemicals do not match")
         if basis == 'mol':
-            return material.mol
+            return material.imol.data 
         elif basis == 'wt':
-            return material.mass
+            return material.mass.data
         else:
             raise ValueError("basis must be either 'mol' or 'wt'")
     else:
@@ -83,12 +96,12 @@ class Reaction:
     Other Parameters
     ----------------
     check_mass_balance=False: bool
-        Whether to assert that mass is not created or destroyed.
+        Whether to check if mass is not created or destroyed.
     correct_mass_balance=False: bool
         Whether to make sure mass is not created or destroyed by varying the 
         reactant stoichiometric coefficient.
     check_atomic_balance=False: bool
-        Whether to assert that stoichiometric balance by atoms cancel out.
+        Whether to check if stoichiometric balance by atoms cancel out.
     correct_atomic_balance=False: bool
         Whether to correct the stoichiometry according to the atomic balance.
     
@@ -156,7 +169,9 @@ class Reaction:
     array([ 60.   ,  15.666, 124.334])
     
     """
+    phases = MaterialIndexer.phases
     __slots__ = ('_basis',
+                 '_phases',
                  '_chemicals',
                  '_X_index', 
                  '_stoichiometry', 
@@ -174,9 +189,19 @@ class Reaction:
             raise ValueError("basis must be either by 'wt' or by 'mol'")
         self.X = X
         chemicals = self._load_chemicals(chemicals)
-        if reaction:            
-            self._stoichiometry = prs.get_stoichiometric_array(reaction, chemicals)
-            self._X_index = self._chemicals.index(reactant)
+        if reaction:
+            phases = xprs.get_phases(reaction)
+            if phases:
+                self._phases = phases = tuple(set(phases))
+                self._stoichiometry = stoichiometry = xprs.get_stoichiometric_array(reaction, phases, chemicals)
+                reactant_index = self._chemicals.index(reactant)
[…]
