### 2020-11-20 977c086
fix bugs related to new phases feature of Reaction objects
