### 2020-11-20 9f86928
fix bug related to new phases feature of Reaction objects
