+++ thermosteam/_phase.py
@@ -7,7 +7,7 @@
 # for license details.
 """
 """
-__all__ = ('Phase', 'LockedPhase', 'NoPhase', 'AnyPhase')
+__all__ = ('Phase', 'LockedPhase', 'NoPhase')
 
 isa = isinstance
 new = object.__new__
@@ -70,13 +70,3 @@ class LockedPhase(Phase):
             raise AttributeError('phase is locked')
         
 NoPhase = LockedPhase(None)
-
-class AnyPhase:
-    
-    def __eq__(self, other):
-        return other in valid_phases
-    
-    def __repr__(self):
-        return "AnyPhase"
-    
-AnyPhase = AnyPhase()
