### 2020-11-20 f27eef0
further work on reactions with phases
