### 2020-11-20 ae49380
fix reaction phase bugs and add examples
