+++ thermosteam/reaction/_reaction.py
@@ -11,26 +11,27 @@ import thermosteam as tmo
 import flexsolve as flx
 from chemicals import elements
 from warnings import warn
+from collections.abc import Sized
 from . import (
     _parse as prs,
     _xparse as xprs,
 )
 from ..utils import chemicals_user
-from .._phase import NoPhase
+from .._phase import NoPhase, phase_tuple
 from ..indexer import ChemicalIndexer, MaterialIndexer
 from ..exceptions import InfeasibleRegion
 import numpy as np
 
 __all__ = ('Reaction', 'ParallelReaction', 'SeriesReaction')
 
-def get_stoichiometric_string(reaction):
-    if reaction.any_phase:
-        return prs.get_stoichiometric_string(reaction._stoichiometry, 
-                                             reaction._chemicals)
+def get_stoichiometric_string(stoichiometry, phases, chemicals):
+    if phases:
+        return xprs.get_stoichiometric_string(stoichiometry,
+                                              phases,
+                                              chemicals)
     else:
-        return xprs.get_stoichiometric_string(reaction._stoichiometry,
-                                              reaction._phases,
-                                              reaction._chemicals)
+        return prs.get_stoichiometric_string(stoichiometry, 
+                                             chemicals)
 
 def react_stream_adiabatically(stream, reaction):
     if not isinstance(stream, tmo.Stream):
@@ -124,7 +125,7 @@ class Reaction:
     >>> reaction.show() # Note that the default basis is by 'mol'
     Reaction (by mol):
      stoichiometry             reactant    X[%]
-     H2O,l -> H2,g + 0.5 O2,g  H2O        70.00
+     H2O,l -> H2,g + 0.5 O2,g  H2O,l       70.00
     >>> feed = tmo.Stream('feed', H2O=100)
     >>> feed.phases = ('g', 'l') # Gas and liquid phases must be available
     >>> reaction(feed) # Call to run reaction on molar flow
@@ -141,7 +142,7 @@ class Reaction:
     >>> reaction.show()
     Reaction (by wt):
      stoichiometry                     reactant    X[%]
-     H2O,l -> 0.112 H2,g + 0.888 O2,g  H2O        70.00
+     H2O,l -> 0.112 H2,g + 0.888 O2,g  H2O,l       70.00
     
     Although we changed the basis, the end result is the same if we pass a 
     stream:
@@ -188,6 +189,7 @@ class Reaction:
     
     def __init__(self, reaction, reactant, X, 
                  chemicals=None, basis='mol', *,
+                 phases=None,
                  check_mass_balance=False,
                  check_atomic_balance=False,
                  correct_atomic_balance=False,
@@ -199,16 +201,14 @@ class Reaction:
         self.X = X
         chemicals = self._load_chemicals(chemicals)
         if reaction:
-            phases = xprs.get_phases(reaction)
+            self._phases = phases = phase_tuple(phases) if phases else xprs.get_phases(reaction)
             if phases:
-                self._phases = phases = phases
                 self._stoichiometry = stoichiometry = xprs.get_stoichiometric_array(reaction, phases, chemicals)
                 reactant_index = self._chemicals.index(reactant)
                 for phase_index, x in enumerate(stoichiometry[:, reactant_index]):
                     if x: break
                 self._X_index = (phase_index, reactant_index)
             else:
-                self._phases = phases
                 self._stoichiometry = prs.get_stoichiometric_array(reaction, chemicals)
                 self._X_index = self._chemicals.index(reactant)
             self._rescale()
@@ -224,11 +224,6 @@ class Reaction:
         else:
             self._stoichiometry = np.zeros(chemicals.size)
[…]
