### 2020-11-21 4f23d74
document PhaseIndexer object
