### 2020-11-21 310cc53
add new reactions with phase change features to docs and fix minor bugs
