+++ thermosteam/reaction/_reaction.py
@@ -126,6 +126,8 @@ class Reaction:
     Reaction (by mol):
      stoichiometry             reactant    X[%]
      H2O,l -> H2,g + 0.5 O2,g  H2O,l       70.00
+    >>> reaction.reactant # The reactant is a tuple of phase and chemical ID
+    ('l', 'H2O')
     >>> feed = tmo.Stream('feed', H2O=100)
     >>> feed.phases = ('g', 'l') # Gas and liquid phases must be available
     >>> reaction(feed) # Call to run reaction on molar flow
@@ -157,7 +159,7 @@ class Reaction:
                          O2   35
                      (l) H2O  30
     
-    If you don't specify the phase of your chemicals, Reaction object can 
+    If chemicals phases are not specified, Reaction objects can 
     react a any single phase Stream object (regardless of phase):
     
     >>> reaction = rxn.Reaction('2H2O -> 2H2 + O2', reactant='H2O', X=0.7)
@@ -379,7 +381,7 @@ class Reaction:
          H2 + 0.5 O2 -> H2O  H2         70.00
         
         >>> reaction(s1) 
-        >>> s1.show() # After non-adiabatic reaction
+        >>> s1.show() # After isothermal reaction
         Stream: s1
          phase: 'g', T: 373.15 K, P: 101325 Pa
          flow (kmol/hr): H2   3
@@ -450,7 +452,7 @@ class Reaction:
         """[str] Reactant associated to conversion."""
         if self._phases:
             phase_index, chemical_index = self._X_index
-            return self._chemicals.IDs[chemical_index] + ',' + self._phases[phase_index]
+            return self._phases[phase_index], self._chemicals.IDs[chemical_index]
         else:
             return self._chemicals.IDs[self._X_index] 
 
@@ -543,10 +545,9 @@ class Reaction:
         Balance glucose fermentation to ethanol:
         
         >>> import thermosteam as tmo
-        >>> from thermosteam import reaction as rxn
         >>> from biorefineries import lipidcane as lc
         >>> tmo.settings.set_thermo(lc.chemicals)
-        >>> fermentation = rxn.Reaction('Glucose + O2 -> Ethanol + CO2',
+        >>> fermentation = tmo.Reaction('Glucose + O2 -> Ethanol + CO2',
         ...                             reactant='Glucose',  X=0.9)
         >>> fermentation.correct_atomic_balance()
         >>> fermentation.show()
@@ -556,18 +557,29 @@ class Reaction:
         
         Balance methane combustion:
             
-        >>> combustion = rxn.Reaction('CH4 + O2 -> Water + CO2',
+        >>> combustion = tmo.Reaction('CH4 + O2 -> Water + CO2',
         ...                           reactant='CH4', X=1)
         >>> combustion.correct_atomic_balance()
         >>> combustion.show()
         Reaction (by mol):
          stoichiometry                reactant    X[%]
          2 O2 + CH4 -> 2 Water + CO2  CH4       100.00
+         
+        Balance electrolysis of water (with chemical phases specified):
+            
+        >>> electrolysis = tmo.Reaction('H2O,l -> H2,g + O2,g',
+        ...                             chemicals=tmo.Chemicals(['H2O', 'H2', 'O2']),
+        ...                             reactant='H2O', X=1)
+        >>> electrolysis.correct_atomic_balance()
+        >>> electrolysis.show()
+        Reaction (by mol):
+         stoichiometry             reactant    X[%]
+         H2O,l -> H2,g + 0.5 O2,g  H2O,l     100.00
         
[…]
