### 2020-11-21 a47e52b
fix documentation details
