### 2020-11-21 BioSTEAMDevelopmentGroup/thermosteam#29 @yoelcortes
Hi Francesco,

I added some cool new features to the Reaction, ParallelReaction, and SeriesReaction objects to be able to handle phase changes:

```python
import thermosteam as tmo
from thermosteam import reaction as rxn

chemicals = tmo.Chemicals(['Water', 'Methane', 'CO2', 'O2', 'N2'])
tmo.settings.set_thermo(chemicals)
combustion = rxn.Reaction('Methane,g + O2,g -> Water,l + CO2,g',
                          reactant='Methane', X=1,
                          correct_atomic_balance=True)

feed = tmo.Stream(Methane=1, O2=2, N2=100, T=373.15, phase='g') # Dilute with nitrogen to keep temperature change low
feed.phases = 'lg'
print('BEFORE ADIABATIC REACTION')
feed.show(N=100)

# React feed molar flow rate
combustion.adiabatic_reaction(feed)

print('AFTER ADIABATIC REACTION')
feed.show(N=100)
```

Output:
```python
BEFORE ADIABATIC REACTION
MultiStream: s1
 phases: ('g', 'l'), T: 373.15 K, P: 101325 Pa
 flow (kmol/hr): (g) Methane  1
                     O2       2
                     N2       100
AFTER ADIABATIC REACTION
MultiStream: s1
 phases: ('g', 'l'), T: 648.24 K, P: 101325 Pa
 flow (kmol/hr): (g) CO2    1
                     N2     100
                 (l) Water  2
```

Note that latent heats will change depending on the temperature, so Reaction objects can only take into account heats of formation (not latent heats). In any case, by making the reaction adiabatic, these latent heats are accounted for in the energy balance. I added a comment in the tutorial that `dH` does not account for latent heats, and a warning in the documentation as well:
* https://thermosteam.readthedocs.io/en/latest/reaction/Reaction.html#thermosteam.reaction.Reaction.dH
* https://thermosteam.readthedocs.io/en/latest/tutorial/Stoichiometric_reactions.html#Single-reaction

I'll go ahead and close the issue, but feel free to reopen if anything comes up.

Thanks!
