### 2020-11-21 BioSTEAMDevelopmentGroup/thermosteam#29 @yoelcortes
@fwitte Almost forgot to mention. I'm not too sure if this interests you, but it is possible to calculate heats of reaction with latent heats as follows:

```python
import thermosteam as tmo
chemicals = tmo.Chemicals(['Water', 'Methane', 'CO2', 'O2', 'N2'])
tmo.settings.set_thermo(chemicals)
combustion = tmo.Reaction('Methane,g + O2,g -> Water,l + CO2,g',
                          reactant='Methane', X=1,
                          correct_atomic_balance=True)
feed = tmo.Stream(Methane=1, O2=2, N2=100, T=298.15, phase='g')
feed.phases = 'lg'
H0 = feed.Hnet
combustion(feed)
Hf = feed.Hnet
print(f'Heat of reaction with latent heats: {Hf - H0:.2f} kJ/hr')
print(f'Higher heating value of methane: {chemicals.Methane.HHV:.2f} J/mol')
```

Output:
```python
Heat of reaction with latent heats: -890590.00 kJ/hr
Higher heating value of methane: -890590.00 J/mol
```

Thanks!
