### 2020-11-21 2587198
add note on how to shadow clone thermosteam to clone within seconds
