### 2020-11-21 BioSTEAMDevelopmentGroup/thermosteam#23 @yoelcortes
Hi Francesco, I found and easy solution on the user-side to clone the repository within seconds and added a note on it in the "README" file and the documentation (https://thermosteam.readthedocs.io/en/latest/Installation.html). The idea is just to clone the last 100 or so commits using "git clone --depth [depth] [url]". I'll go ahead an close this issue, but let me know if you have any questions!

Thanks,
