+++ docs/paper/paper.bib
@@ -45,6 +45,31 @@
   url = {https://github.com/CalebBell/chemicals}
 }
 
+@misc{CoolProp,
+  author = {{​Ian Bell and Contributors}},
+  title = {CoolProp: Thermophysical properties for the masses},
+  year = {2020},
+  publisher = {​GitHub},
+  journal = {​GitHub repository},
+  url = {https://github.com/CoolProp/CoolProp}
+}
+
+@software{Cantera,
+  author       = {Goodwin, David G and
+                  Speth, Raymond L and
+                  Moffat, Harry K and
+                  Weber, Bryan W},
+  title        = {{Cantera: An Object-oriented Software Toolkit for 
+                   Chemical Kinetics, Thermodynamics, and Transport
+                   Processes}},
+  month        = aug,
+  year         = 2018,
+  publisher    = {Zenodo},
+  version      = {2.4.0},
+  doi          = {10.5281/zenodo.1174508},
+  url          = {https://doi.org/10.5281/zenodo.1174508}
+}
+
 @incollection{Perry,
 	address = {New York, NY},
 	edition = {9th},
