+++ docs/paper/paper.md
@@ -24,39 +24,50 @@ bibliography: paper.bib
 
 # Summary
 
-The design and simulation of unit operations in a chemical process often 
-requires the estimation of mixture properties and phase equilibrium (e.g. 
-vapor-liquid and liquid-liquid equilibrium) to correctly estimate material and 
-energy balances, as well as the unit operation design requirements to achieve 
-a set of design specifications. For example, estimating both phase equilibrium
-and fluid viscosities, densities, and surface tensions are required to design a 
-distillation column that can achieve a specified recovery of chemicals 
-[@Perry]. The overarching goal of thermosteam is to enable the rigorous design 
-and simulation of unit operations by creating thermodynamic property packages
-from both user-defined chemicals and databanks from the `chemicals` library [@chemicals], an open-source compilation of data and functions for the estimation of thermodynamic and transport properties for both pure chemicals and mixtures.
+`Thermosteam` is a standalone thermodynamic engine capable of solving mass and 
+energy balances, estimating mixture properties, solving thermodynamic phase 
+equilibria, and modeling stoichiometric reactions. Before getting started with 
+thermodynamic calculations, `Thermosteam` first creates a thermodynamic property 
+package that defines the models used to estimate mixture properties and phase 
+equilibria from both user-defined chemicals and databanks from the `chemicals` 
+library [@chemicals], an open-source compilation of data and functions for the 
+estimation of pure component chemical and mixture properties. Roughly 20,000 chemicals 
+with temperature- and pressure-dependent property data are included in the 
+`chemicals` library. `Thermosteam` builds upon `chemicals` with a robust and 
+flexible framework that facilitates the creation of property packages. 
+Users may easily define new models and pseudo-chemicals that are compatible with
+all of `Thermosteam`'s features. 
 
-Roughly 20,000 chemicals with temperature- and pressure-dependent property data are
-included in the `chemicals` library. Thermosteam builds upon `chemicals` with a robust and flexible framework that facilitates the creation of property packages. Its extendable framework allows for easy integration of new models for estimating pure component properties, thermodynamic equilibrium coefficients, and mixture properties. 
-The Biorefinery Simulation and Techno-Economic Analysis Modules (BioSTEAM) 
-has adopted thermosteam as its premier thermodynamic engine [@BioSTEAM].
-Published biorefinery designs modeled in BioSTEAM implement property 
-packages created with thermosteam [@Bioindustrial-Park], including a cornstover 
-biorefinery for the production of cellulosic ethanol, a lipid-cane biorefinery 
-for the co-production of ethanol and biodiesel, a sugarcane biorefinery
-for the production of bioethanol, and a wheatstraw biorefinery for the production
-of cellulosic ethanol [@BioSTEAM][@Sanchis].
+The overarching goal of `Thermosteam` is to enable the rigorous design and 
+simulation of chemical production processes, whereby low value feedstocks are
+converted to high value products via chemical reactions and thermodynamic-driven
+separations. For example, modeling the separation of volatile chemicals from 
+heavier ones in a distillation column (e.g. distilling ethanol from water),
+requires vapor-liquid phase equilibrium calculations to predict how well volatile 
+chemicals selectively partition into the vapor phase. Additionally, fluid 
+viscosities, densities, and surface tensions are required to appropriately design 
+a distillation column that can achieve a specified recovery of chemicals 
+[@Perry]. The Biorefinery Simulation and Techno-Economic Analysis 
+Modules (BioSTEAM) — capable of modeling distillation columns, reactors,
+heat exchangers, and other unit operations — has adopted `Thermosteam` 
+as its premier thermodynamic engine [@BioSTEAM]. Published designs modeled in 
[…]
