### 2020-11-22 f7db43e
minor enhancements to paper
