### 2020-11-22 #2814 @yoelcortes
Hi @decaluwe and @fwitte,

I have taken care of the issues on Github and made edits to the manuscript. I'd like to address each of your comments in detail here and note the changes I made:

Professor DeCaluwe:

1.   Depending on one's definition of "a diverse, non-specialist audience," the intro could probably be written a little bit broader. For example, phrases like "unit operations" and even "chemical process" might be opaque for somebody in materials science or astronomy. I'd suggest something like: (i) end the first sentence right before the parenthetical, (ii) describe some common features of a "unit operation in a chemical process" (e.g. something along the lines of "fluid flows enter a reactor wherein chemical reactions occur, consuming or producing chemical species and thermal energy...") and then (iii) give some examples/categories of the specific types of properties needed for evaluating the design of a chemical process. It might add a sentence or two, but should be clearer for a non-expert. The distillation column example provides a good specific example, but something helping bridge the very general first sentence and this more specific example would be helpful:
  * This was a very helpful point. I changed the first paragraph to focus on introducing thermosteam's central features, while the 2nd paragraph focuses on defining the overarching goal and giving examples of its usage to a broader audience. The first sentences of the second paragraph now reads:
     *  "The overarching goal of Thermosteam is to enable the rigorous design and simulation of
chemical production processes, whereby low value feedstocks are converted to high value products
via chemical reactions and thermodynamic-driven separations. For example, modeling
the separation of volatile chemicals from heavier ones in a distillation column (e.g. distilling
ethanol from water), requires vapor-liquid phase equilibrium calculations to predict how
well volatile chemicals selectively partition into the vapor phase. Additionally, fluid viscosities,
densities, and surface tensions are required to appropriately design a distillation column
that can achieve a specified recovery of chemicals (Green, 2018). The Biorefinery Simulation
and Techno-Economic Analysis Modules (BioSTEAM) — capable of modeling distillation
columns, reactors, heat exchangers, and other unit operations — has adopted Thermosteam
as its premier thermodynamic engine (Cortés-Peña et al., 2020)."

2.  State of the field: There is no documenting of or comparison to current capabilities in the field. Would software like EES, CoolProp, RefProp, Cantera, or Chemkin be valid points of comparison? These packages might have similar functionality, but I do not know if they are "commonly used" in the field (which is what the review guidelines specifically ask), so I'd rely on your expertise, obviously. What approaches are commonly used in your field?
  * I added a paragraph that compares CoolProp and Cantera to Thermosteam. I limited the scope to only open-source software. Usually Aspen and other process simulators are used to solve mass and energy balances and phase equilibrium in Chemical engineering while REFPROP may be used when very accurate estimations are needed for chemical properties. However, it's not Thermosteam's goal to be a process simulator or to provide state-of-the-art estimations of thermophysical properties. Thus, comparing thermosteam to just CoolProp and Cantera may be more beneficial. The new paragraph reads as follows:
    * "Various other open-source libraries in Python have comparable capabilities to
`Thermosteam`: most notably `Cantera` and `CoolProp`. `Cantera` is
a collection of software tools capable of modeling kinetic reactions,
thermodynamic equilibrium, and chemical mixture properties [@Cantera]. `Cantera`'s
built-in chemicals are limited to 8, but new chemicals can be defined by users
with flexibility on the amount of detail. `Thermosteam` has yet to implement any
features on kinetic reaction networks, but exposes a larger set of roughly 20,000
built-in chemicals from the `chemicals` library. `CoolProp` offers fast and accurate
thermodynamic and transport properties for 122 chemical components, and can
estimate thermodynamic phase equilibrium and mixture properties [@CoolProp].
`CoolProp` also offers an interface to the NIST REFPROP software, which is
considered the gold standard in thermophysical properties [@REFPROP]. It is
within `Thermosteam`'s roadmap to use `CoolProp` as part of it's default, built-in
models. While `CoolProp` focuses on thermophysical chemical properties,
`Thermosteam` also includes mass and energy balances and stoichiometric reactions
as one of its central features."

Professor Witte:

1. The paper highlights the most important features of the software, but (as already written in the docs issue) I do miss the capability, that generic chemical reactions can be computed with your software. I think this is a very important feature of your software.
  * Thanks! I highlighted this feature in the following sentences:
    * [first sentence of the 1st paragraph] "`Thermosteam` is a standalone thermodynamic engine capable of solving mass and energy balances, estimating mixture properties, solving thermodynamic phase
equilibria, and modeling stoichiometric reactions."
    * [last sentence of the 5th paragraph] "While `CoolProp` focuses on thermophysical chemical properties,
`Thermosteam` also includes mass and energy balances and stoichiometric reactions
as one of its central features."
2. A statement regarding the main advantages compared to software in the field (like @decaluwe pointed out) would be very good to emphasize the value of thermosteam, too.
  * Thanks for noting this. Please see my reply to Professor DeCaluwe 2nd comment.
3. I am not a legal expert, is it necessary to include the disclaimer twice. Can it not be handled in a single sentence? Maybe something like "Any opinions, findings, and conclusions or recommendations expressed in this publication are those of the author and necessarily reflect the views of the funders and sponsors."?
  * I have contacted my institution about this and they prefer to keep this exact wording, unfortunately.
5. Are these [@BioSTEAM][@Sanchis] meant to be two individual citations? You can connect them with [@BioSTEAM; @Sanchis]
  * Fixed!
