### 2020-11-23 8e1edb8
add documentation of math operations with reactions
