### 2020-11-23 c66074c
reactants property is now tuple instead of list; better examples for material balance function
