### 2020-11-23 c03d6be
add documentation, reduce number of siginficant digits in indexers and multi-stream objects, and update version
