+++ thermosteam/separations.py
@@ -231,6 +231,8 @@ def phase_split(feed, outlets):
         
     Examples
     --------
+    Split gas and liquid phases to streams:
+    
     >>> import thermosteam as tmo
     >>> tmo.settings.set_thermo(['Water', 'Ethanol'], cache=True)
     >>> feed = tmo.Stream('feed', Water=10, Ethanol=10)
@@ -249,6 +251,13 @@ def phase_split(feed, outlets):
      phase: 'l', T: 353.88 K, P: 101325 Pa
      flow (kmol/hr): Water    6.14
                      Ethanol  3.86
+    
+    Note that the number of phases in the feed should be equal to the number of 
+    outlets:
+        
+    >>> tmo.separations.phase_split(feed, [vapor])
+    Traceback (most recent call last):
+    RuntimeError: number of phases in feed must be equal to the number of outlets
     
     """
     phases = feed.phases
@@ -460,7 +469,7 @@ def lle(feed, top, bottom, top_chemical=None, efficiency=1.0, multi_stream=None)
     bottom : Stream
         Bottom fluid.
     top_chemical : str, optional
-        Identifier of chemical that will be favored in the "liquid" phase.
+        Identifier of chemical that will be favored in the top fluid.
     efficiency=1. : float,
         Fraction of feed in liquid-liquid equilibrium.
         The rest of the feed is divided equally between phases
@@ -469,6 +478,8 @@ def lle(feed, top, bottom, top_chemical=None, efficiency=1.0, multi_stream=None)
     
     Examples
     --------
+    Perform liquid-liquid equilibrium around water and octanol and split the phases:
+    
     >>> import thermosteam as tmo
     >>> tmo.settings.set_thermo(['Water', 'Ethanol', 'Octanol'], cache=True)
     >>> feed = tmo.Stream('feed', Water=20, Octanol=20, Ethanol=1)
@@ -487,6 +498,25 @@ def lle(feed, top, bottom, top_chemical=None, efficiency=1.0, multi_stream=None)
      flow (kmol/hr): Water    16.5
                      Ethanol  0.139
                      Octanol  0.00409
+    
+    Assume that 1% of the feed is not in equilibrium (possibly due to poor mixing):
+        
+    >>> import thermosteam as tmo
+    >>> tmo.settings.set_thermo(['Water', 'Ethanol', 'Octanol'], cache=True)
+    >>> feed = tmo.Stream('feed', Water=20, Octanol=20, Ethanol=1)
+    >>> top = tmo.Stream('top')
+    >>> bottom = tmo.Stream('bottom')
+    >>> ms = tmo.MultiStream('ms', phases='lL') # Store flow rate data here as well
+    >>> tmo.separations.lle(feed, top, bottom, efficiency=0.99, multi_stream=ms)
+    >>> ms.show()
+    MultiStream: ms
+     phases: ('L', 'l'), T: 298.15 K, P: 101325 Pa
+     flow (kmol/hr): (L) Water    3.55
+                         Ethanol  0.861
+                         Octanol  20
+                     (l) Water    16.5
+                         Ethanol  0.139
+                         Octanol  0.00408
     
     """
     if multi_stream:
@@ -545,6 +575,9 @@ def vle(feed, vap, liq, T=None, P=None, V=None, Q=None, x=None, y=None,
     
     Examples
     --------
+    Perform vapor-liquid equilibrium on water and ethanol and split phases
+    to vapor and liquid streams:
+    
     >>> import thermosteam as tmo
     >>> tmo.settings.set_thermo(['Water', 'Ethanol'], cache=True)
     >>> feed = tmo.Stream('feed', Water=20, Ethanol=20)
[…]
