### 2020-11-23 BioSTEAMDevelopmentGroup/thermosteam#29 @fwitte
Hi Yoel,
that is a very good addition to thermosteam, I think!

But, I actually have one more question: There seems to be a little mismatch between the lower heating value and the heat of reaction when water is in gaseous state. I cannot really spot the origin of that. Maybe an issue in the reference point definition?

```python

chemicals = tmo.Chemicals(['Water', 'Methane', 'CO2', 'O2', 'N2'])
tmo.settings.set_thermo(chemicals)
combustion = tmo.Reaction('Methane + O2 -> Water + CO2',
                          reactant='Methane', X=1,
                          correct_atomic_balance=True)
feed = tmo.Stream(Methane=1, O2=2, N2=100, T=350.15, phase='g')
H0 = feed.Hnet
combustion(feed)
Hf = feed.Hnet
print(f'Heat of reaction with latent heats: {Hf - H0:.2f} kJ/hr')
print(f'Lower heating value of methane: {chemicals.Methane.LHV:.2f} J/mol')
```

```bash
Heat of reaction with latent heats: -797156.83 kJ/hr
Lower heating value of methane: -802567.01 J/mol
```
It is more obvious with the electrolysis of water: The heat of reaction with liquid water is correct, the heat of reaction with gaseous water is slightly below the expected value (which should be the enthalpy of formation of gaseous water, so about -241.8 kJ/mol, or?).
```python

chemicals = tmo.Chemicals(['H2O', 'H2', 'O2'], cache=True)
tmo.settings.set_thermo(chemicals)
reaction = rxn.Reaction('2H2O,l -> 2H2,g + O2,g', reactant='H2O', X=1)
reaction.show() # Note that the default basis is by 'mol'
reaction.reactant # The reactant is a tuple of phase and chemical ID
feed = tmo.Stream('feed', H2O=1)
H0 = feed.Hnet
feed.phases = ('g', 'l') # Gas and liquid phases must be available
reaction(feed) # Call to run reaction on molar flow
Hf = feed.Hnet
print(f'Heat of reaction with latent heats: {Hf - H0:.2f} kJ/hr')

reaction = rxn.Reaction('2H2O -> 2H2 + O2', reactant='H2O', X=1)
reaction.show() # Note that the default basis is by 'mol'
reaction.reactant # The reactant is a tuple of phase and chemical ID
feed = tmo.Stream('feed', H2O=1, T=298.15, phase='g')
H0 = feed.Hnet
reaction(feed) # Call to run reaction on molar flow
Hf = feed.Hnet
print(f'Heat of reaction with latent heats: {Hf - H0:.2f} kJ/hr')
```

```bash
Reaction (by mol):
 stoichiometry             reactant    X[%]
 H2O,l -> H2,g + 0.5 O2,g  H2O,l     100.00
Heat of reaction with latent heats: 285825.00 kJ/hr
Reaction (by mol):
 stoichiometry       reactant    X[%]
 H2O -> H2 + 0.5 O2  H2O       100.00
Heat of reaction with latent heats: 239365.21 kJ/hr
```
The difference between heat of reaction and lower heating value in the combustion reaction of methane is actually very close to the offset in the gaseous water.

Maybe you can help me spot the issue.
Thank you and best regards
Francesco
