### 2020-11-23 #2814 @fwitte
Hi Yoel,
thank you very much for taking your time and carefully addressing all of my comments and issues. Regarding the paper, I really think the second last paragraph does a very good job of highlighting thermosteam's advantages and differentiation.
For me, there is only one more comment to be addressed (regarding the heat of formation of gasoues water in BioSTEAMDevelopmentGroup/thermosteam#29).
Best regards, have a nice week!
Francesco
