### 2020-11-23 BioSTEAMDevelopmentGroup/thermosteam#29 @yoelcortes
Hi Fracesco,

This one is easy to spot for me, but I know its hard to find at first sight. The heat capacities of reactants and products are different. Thus, this adds **changes in sensible heats** (in addition to the latent heats and heats of formation). Note that in my example, I used a temperature of 298.15 K. At this temperature, the enthalpy of chemicals at their reference state is 0 (since this is the reference point). That is why it worked before. Perhaps this threw you off to generalize, sorry!

You might ask yourself why is the electrolysis of water in the gas phase appear to be off as well? This is because the enthalpy of gaseous water at 298.15 K is not at the reference state (the reference state is a liquid). Note that heats of vaporization change depending on the temperature (another way to view it is to go from 25C to 100C as liquid water, boil, then go from 100C to 25 as steam). Perhaps this quick example may help:

```python
import thermosteam as tmo
Water = tmo.Chemical('Water')
print(f"T       H_vap")
for i in range(20):
    T = 298.15 + 5*i
    H_liq = Water.H('l', T)
    H_gas = Water.H('g', T)
    H_vap = H_gas - H_liq
    print(f"{T:.2f}, {H_vap:.2f}") # Values may differ depending what models you use. Eventually I hope to implement IAPWS models...
```
Output [EDITED after bug in gas enthalpy was fixed]:
```
T       H_vap
298.15, 43927.16
303.15, 43718.68
308.15, 43510.46
313.15, 43302.43
318.15, 43094.50
323.15, 42886.59
328.15, 42678.65
333.15, 42470.61
338.15, 42262.42
343.15, 42054.05
348.15, 41845.45
353.15, 41636.60
358.15, 41427.48
363.15, 41218.08
368.15, 41008.39
373.15, 40798.42
378.15, 40588.18
383.15, 40377.68
388.15, 40166.96
393.15, 39956.01
```

If you are looking to get the exact value for HHVs and LHV (without taking into account sensible heats), it is best to perform these calculations separately. Thermosteam uses this function from the `chemicals` library to compute HHV and LHV:

```python
from chemicals import combustion_data, Hfg, CAS_from_any
CAS = CAS_from_any('Methane')
cd = combustion_data('CH4', Hf=Hfg(CAS)) # If Hf is not passed, combustion is estimated using Dulong's formula (not accurate here)
print(cd)
print(cd.LHV) # A property
```
Output:
```
CombustionData(stoichiometry={'CO2': 1, 'O2': -2.0, 'H2O': 2.0}, HHV=-890590.0, Hf=-74534.0, MW=16.04246)
-802567.008
```
I hope this helps,
Let me know if you have any more questions!
