### 2020-11-23 BioSTEAMDevelopmentGroup/thermosteam#29 @yoelcortes
@fwitte, I just made a correction to the combustion_data code block. I forgot to pass Hf (heat of formation of the chemical).

Thanks,
Sorry for the confusion!
