### 2020-11-23 f0ca08f
add additional reaction tests for better coverage
