+++ thermosteam/reaction/_reaction.py
@@ -116,17 +116,16 @@ class Reaction:
     Warning
     -------
     Negative conversions and conversions above 1.0 are fair game (allowed), but
-    may lead to infeasible values when reacting a stream.
+    may lead to odd/infeasible values when reacting a stream.
     
     Examples
     --------
     Electrolysis of water to molecular hydrogen and oxygen:
     
     >>> import thermosteam as tmo
-    >>> import thermosteam.reaction as rxn
     >>> chemicals = tmo.Chemicals(['H2O', 'H2', 'O2'], cache=True)
     >>> tmo.settings.set_thermo(chemicals)
-    >>> reaction = rxn.Reaction('2H2O,l -> 2H2,g + O2,g', reactant='H2O', X=0.7)
+    >>> reaction = tmo.Reaction('2H2O,l -> 2H2,g + O2,g', reactant='H2O', X=0.7)
     >>> reaction.show() # Note that the default basis is by 'mol'
     Reaction (by mol):
      stoichiometry             reactant    X[%]
@@ -167,7 +166,7 @@ class Reaction:
     If chemicals phases are not specified, Reaction objects can 
     react a any single phase Stream object (regardless of phase):
     
-    >>> reaction = rxn.Reaction('2H2O -> 2H2 + O2', reactant='H2O', X=0.7)
+    >>> reaction = tmo.Reaction('2H2O -> 2H2 + O2', reactant='H2O', X=0.7)
     >>> feed = tmo.Stream('feed', H2O=100, phase='g')
     >>> reaction(feed) 
     >>> feed.show() 
@@ -400,10 +399,9 @@ class Reaction:
         heat of reaction:
         
         >>> import thermosteam as tmo
-        >>> import thermosteam.reaction as rxn
         >>> chemicals = tmo.Chemicals(['H2', 'O2', 'H2O'], cache=True)
         >>> tmo.settings.set_thermo(chemicals)
-        >>> reaction = rxn.Reaction('2H2 + O2 -> 2H2O', reactant='H2', X=0.7)
+        >>> reaction = tmo.Reaction('2H2 + O2 -> 2H2O', reactant='H2', X=0.7)
         >>> s1 = tmo.Stream('s1', H2=10, O2=20, H2O=1000, T=373.15, phase='g')
         >>> s2 = tmo.Stream('s2')
         >>> s2.copy_like(s1) # s1 and s2 are the same
@@ -901,15 +899,16 @@ class ParallelReaction(ReactionSet):
     
     Examples
     --------
+    Run two reactions in parallel:
+    
     >>> import thermosteam as tmo
-    >>> import thermosteam.reaction as rxn
     >>> chemicals = tmo.Chemicals(['H2', 'Ethanol', 'CH4', 'O2', 'CO2', 'H2O'], cache=True)
     >>> tmo.settings.set_thermo(chemicals)
     >>> kwargs = dict(phases='lg', correct_atomic_balance=True)
-    >>> reaction = rxn.ParallelReaction([
+    >>> reaction = tmo.ParallelReaction([
     ...    #            Reaction definition                    Reactant             Conversion
-    ...    rxn.Reaction('H2,g + O2,g -> 2H2O,g',               reactant='H2',       X=0.7, **kwargs),
-    ...    rxn.Reaction('Ethanol,l + O2,g -> CO2,g + 2H2O,g',  reactant='Ethanol',  X=0.1, **kwargs)
+    ...    tmo.Reaction('H2,g + O2,g -> 2H2O,g',               reactant='H2',       X=0.7, **kwargs),
+    ...    tmo.Reaction('Ethanol,l + O2,g -> CO2,g + 2H2O,g',  reactant='Ethanol',  X=0.1, **kwargs)
     ... ])
[…]
