### 2020-11-23 966dff8
add tests for models and ignore coverage for plots
