### 2020-11-23 b35331e
add lots of documentation and examples to push coverage to 90%
