### 2020-11-24 BioSTEAMDevelopmentGroup/thermosteam#29 @yoelcortes
Hi Francesco,

I went through the calculations to be able to show that everything balances out, and I found a bug in the calculation of the gas phase enthalpy. Thanks for looking at everything in detail!

Here is the demonstration to show that everything balances out by removing the sensible heats. Please pull my changes before you try the example.

Methane example
```python
import thermosteam as tmo
chemicals = Water, Methane, CO2, O2, N2 = tmo.Chemicals(['Water', 'Methane', 'CO2', 'O2', 'N2'])
Water.H.g.Hvap_Tb = 44011.496 # Depending on the model, this value may be different.
tmo.settings.set_thermo(chemicals)
combustion = tmo.Reaction('Methane + O2 -> Water + CO2',
                          reactant='Methane', X=1,
                          correct_atomic_balance=True)
Tref = 298.15
Tb = Water.Tb
feed = tmo.Stream(Methane=1, O2=2, T=Tb, phase='g')
H0 = feed.Hnet - Methane.Cn.g.integrate_by_T(Tref, Tb) - 2 * O2.Cn.g.integrate_by_T(Tref, Tb)
combustion(feed)
Hf = feed.Hnet - 2 * Water.Cn.l.integrate_by_T(Tref, Tb) - CO2.Cn.g.integrate_by_T(Tref, Tb)
print(f'Heat of reaction with latent heats minus sensible heats: {Hf - H0:.2f} kJ/hr')
print(f'Lower heating value of methane: {chemicals.Methane.LHV:.2f} J/mol')
```
```
Heat of reaction with latent heats minus sensible heats: -802567.01 kJ/hr
Lower heating value of methane: -802567.01 J/mol
```

Electrolysis example:
```python
import thermosteam as tmo
H2O, H2, O2 = chemicals = tmo.Chemicals(['H2O', 'H2', 'O2'], cache=True)
H2O.H.g.Hvap_Tb = 44011.496
tmo.settings.set_thermo(chemicals)
reaction = tmo.Reaction('2H2O,l -> 2H2,g + O2,g', reactant='H2O', X=1)
Tref = 298.15
feed = tmo.Stream('feed', H2O=1)
H0 = feed.Hnet
feed.phases = ('g', 'l') # Gas and liquid phases must be available
reaction(feed) # Call to run reaction on molar flow
Hf = feed.Hnet
print(f'Heat of reaction: {Hf - H0:.2f} kJ/hr')
print(f'HHV: {H2.HHV:.2f} kJ/hr')
reaction = tmo.Reaction('2H2O -> 2H2 + O2', reactant='H2O', X=1)
feed = tmo.Stream('feed', H2O=1, T=Tref, phase='g')
H0 = feed.Hnet - H2O.Cn.l.integrate_by_T(Tref, H2O.Tb) - H2O.Cn.g.integrate_by_T(H2O.Tb, Tref)
reaction(feed) # Call to run reaction on molar flow
Hf = feed.Hnet
print(f'Heat of reaction: {Hf - H0:.2f} kJ/hr')
print(f'LHV: {H2.LHV:.2f} kJ/hr')
```
```
Heat of reaction: 285825.00 kJ/hr
HHV: -285825.00 kJ/hr
Heat of reaction: 241813.50 kJ/hr
LHV: -241813.50 kJ/hr
```

Thanks again,
Have a nice day!
