+++ setup.py
@@ -11,7 +11,7 @@ setup(
     name='thermosteam',
     packages=['thermosteam'],
     license='MIT',
-    version='0.22.3',
+    version='0.22.4',
     description="BioSTEAM's Premier Thermodynamic Engine",
     long_description=open('README.rst').read(),
     author='Yoel Cortes-Pena',
@@ -20,7 +20,7 @@ setup(
                       'chemicals>=0.1.5',
                       'IPython>=7.9.0', 
                       'colorpalette>=0.3.1', 
-                      'biosteam>=2.21.7',
+                      'biosteam>=2.21.8',
                       'pandas>=0.25.2', 
                       'matplotlib>=3.1.1',
                       'numpy>=1.18.1', 
