+++ thermosteam/__init__.py
@@ -51,4 +51,4 @@ pd.set_option('display.max_columns', 10)
 pd.set_option('max_colwidth', 35)
 del np, pd
 
-__version__ = "0.22.3"
+__version__ = "0.22.4"
