### 2020-11-24 BioSTEAMDevelopmentGroup/thermosteam#29 @fwitte
Hi Yoel,
great, thank you for taking your time with this. Honestly, I was questioning, if I did implement the combustion chamber balance over at https://github.com/oemof/tespy incorrectly. I am somewhat relieved now :D.
Thanks again, have a nice week!
Francesco
