### 2020-11-24 a79779c
fix gas phase entropy bug and add tests
