+++ thermosteam/free_energy.py
@@ -53,7 +53,7 @@ def Gas_Enthalpy_Ref_Gas(T, Cn_g, T_ref, H_ref):
 
 @functor(var='H.g')
 def Gas_Enthalpy_Ref_Liquid(T, Cn_g, H_int_T_ref_to_Tb_l, Hvap_Tb, 
-                            T_ref, H_ref, Tb):
+                            Tb, H_ref):
     return H_ref + H_int_T_ref_to_Tb_l + Hvap_Tb + Cn_g.integrate_by_T(Tb, T)
 
 @functor(var='H.g')
@@ -109,8 +109,8 @@ def Gas_Entropy_Ref_Gas(T, P, Cn_g, T_ref, P_ref, S0):
 
 @functor(var='S.g')
 def Gas_Entropy_Ref_Liquid(T, P, Cn_g, S_int_T_ref_to_Tb_l, Svap_Tb,
-                           T_ref, P_ref, S0):
-    return S0 + S_int_T_ref_to_Tb_l + Svap_Tb + Cn_g.integrate_by_T_over_T(T_ref, T) - R*log(P/P_ref)
+                           Tb, P_ref, S0):
+    return S0 + S_int_T_ref_to_Tb_l + Svap_Tb + Cn_g.integrate_by_T_over_T(Tb, T) - R*log(P/P_ref)
 
 @functor(var='S.g')
 def Gas_Entropy_Ref_Solid(T, P, Cn_g, S_int_T_ref_to_Tm_s, Sfus,
