### 2020-11-24 #2814 @decaluwe
Many thanks, @yoelcortes for your thorough attention, here.

1. Paper: this is really well one.  One very minor edit: in the 2nd-to-last paragraph, "it's default, built-in models" should be "its default..." (no apostrophe).

2. Unfortunately I now get 5 tests failing:

______________________________________________________________________________ test_sugarcane ______________________________________________________________________________

    def test_sugarcane():
        from biorefineries import sugarcane as sc
        sc.load()
        units = UnitGroup('Biorefinery', sc.sugarcane_tea.units)
>       assert np.allclose(sc.sugarcane_tea.IRR, 0.10276414072162979, rtol=1e-2)
E       AssertionError: assert False
E        +  where False = <function allclose at 0x7f7ffd234290>(0.106440582162112, 0.10276414072162979, rtol=0.01)
E        +    where <function allclose at 0x7f7ffd234290> = np.allclose
E        +    and   0.106440582162112 = <ConventionalEthanolTEA: sugarcane_sys>.IRR
E        +      where <ConventionalEthanolTEA: sugarcane_sys> = <module 'biorefineries.sugarcane' from '/Users/decaluwe/.local/lib/python3.7/site-packages/biorefineries/sugarcane/__init__.py'>.sugarcane_tea

../../../.local/lib/python3.7/site-packages/biorefineries/tests/test_biorefineries.py:18: AssertionError
______________________________________________________________________________ test_lipidcane ______________________________________________________________________________

    def test_lipidcane():
        from biorefineries import lipidcane as lc
        lc.load()
        units = UnitGroup('Biorefinery', lc.lipidcane_tea.units)
        assert np.allclose(lc.lipidcane_tea.IRR, 0.17804673529883677, rtol=1e-2)
        assert np.allclose(lc.lipidcane_tea.sales, 102369101.03579773, rtol=1e-2)
        assert np.allclose(lc.lipidcane_tea.material_cost, 61761912.9045123, rtol=1e-2)
>       assert np.allclose(lc.lipidcane_tea.installed_equipment_cost, 222127644.25250342, rtol=1e-2)
E       AssertionError: assert False
E        +  where False = <function allclose at 0x7f7ffd234290>(224875041.54906386, 222127644.25250342, rtol=0.01)
E        +    where <function allclose at 0x7f7ffd234290> = np.allclose
E        +    and   224875041.54906386 = <ConventionalEthanolTEA: lipidcane_sys>.installed_equipment_cost
E        +      where <ConventionalEthanolTEA: lipidcane_sys> = <module 'biorefineries.lipidcane' from '/Users/decaluwe/.local/lib/python3.7/site-packages/biorefineries/lipidcane/__init__.py'>.lipidcane_tea

../../../.local/lib/python3.7/site-packages/biorefineries/tests/test_biorefineries.py:35: AssertionError
_____________________________________________________________________________ test_cornstover ______________________________________________________________________________

    def test_cornstover():
        from biorefineries import cornstover as cs
        cs.load()
        MESP = cs.cornstover_tea.solve_price(cs.ethanol)
        units = UnitGroup('Biorefinery', cs.cornstover_tea.units)
>       assert np.allclose(MESP, 0.7383610895500932, rtol=1e-2)
E       assert False
E        +  where False = <function allclose at 0x7f7ffd234290>(0.7308776819336283, 0.7383610895500932, rtol=0.01)
E        +    where <function allclose at 0x7f7ffd234290> = np.allclose

../../../.local/lib/python3.7/site-packages/biorefineries/tests/test_biorefineries.py:47: AssertionError
________________________________________________________________________________ test_LAOs _________________________________________________________________________________

    def g():
>       try: f()

tests/test_biorefineries.py:15:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

    def test_LAOs():
        from biorefineries import LAOs as laos
        laos.load()
        MPSP = laos.get_LAOs_MPSP()
        units = UnitGroup('Biorefinery', laos.LAOs_tea.units)
>       assert np.allclose(MPSP, 1226.0016718824597, rtol=1e-2)
E       assert False
E        +  where False = <function allclose at 0x7f7ffd234290>(1206.1272124746877, 1226.0016718824597, rtol=0.01)
E        +    where <function allclose at 0x7f7ffd234290> = np.allclose

../../../.local/lib/python3.7/site-packages/biorefineries/tests/test_biorefineries.py:62: AssertionError

_____________________________________________________________ [doctest] thermosteam.separations.MultiStageLLE ______________________________________________________________
859     >>> solvent = tmo.Stream('solvent', Octanol=500)
860     >>> stages = tmo.separations.MultiStageLLE(N_stages, feed, solvent)
861     >>> stages.simulate_multi_stage_lle_without_side_draws()
862     >>> stages.raffinate.show()
863     Stream:
864      phase: 'l', T: 298.15 K, P: 101325 Pa
865      flow (kmol/hr): Water     413
866                      Methanol  8.4
867                      Octanol   0.103
868     >>> stages.extract.show()
Differences (unified diff with -expected +actual):
    @@ -1,5 +1,5 @@
     Stream:
      phase: 'L', T: 298.15 K, P: 101325 Pa
    - flow (kmol/hr): Water     87.1
    + flow (kmol/hr): Water     87.2
                      Methanol  41.6
                      Octanol   500

/Users/decaluwe/OneDrive - Colorado School of Mines/github/thermosteam/thermosteam/separations.py:868: DocTestFailure

If it would be helpful for you, happy to add this as an issue at the repo.
