### 2020-11-24 d114582
reduce sigfigs for doctest
