### 2020-11-24 2ef258d
reduce sigfigs for doctest
