### 2020-11-24 #2814 @yoelcortes
Professor DeCaluwe,

Thanks for testing! The tests concerning `biorefineries` did not pass because thermosteam requires a new version of the `biorefineries` library. Regarding the multi-stage liquid-liquid extraction test, I fixed it by reducing the number of significant digits for the test. This is necessary because Python itself is compiled by a compiler. Some math calculations (exp, sin, log, power, etc...) are not bit-for-bit exact between different compilers (and this is not a requirement for them unfortunately), so the end results will always differ.

To get passing tests, please pull my changes on github and reinstall thermosteam (this will update the biorefineries package which is used to run some tests).

Regarding the paper, good catch. I removed "default," from the sentence. Now it reads "It is within `Thermosteam`'s roadmap to use `CoolProp` as part of it's built-in models".

Thanks,
Have a nice day!
