### 2020-11-24 #2814 @decaluwe
Many thanks, @yoelcortes -- all tests now pass.

All of my issues have been resolved - looks great!

@kyleniemeyer LMK if anything else is needed.

Thanks all, for involving me.  it was enjoyable!

Best,
Steven
