### 2020-11-24 #2814 @kyleniemeyer
@yoelcortes one quick suggestion: for tests involving floating-point numbers, you might consider using something like NumPy's [`assert_allclose`](https://numpy.org/doc/stable/reference/generated/numpy.testing.assert_allclose.html), which allows you to test that floating-point values are nearly equal based on a relative and absolute tolerance.
