### 2020-11-24 #2814 @fwitte
> "It is within `Thermosteam`'s roadmap to use `CoolProp` as part of it's built-in models"

Shouldn't it be its and not it's (as it is)?
