### 2020-11-24 #2814 @yoelcortes
@kyleniemeyer, thanks for the suggestion! `assert_allclose` is already used in many tests. For the documentation examples (using doctests), I simply display a lower amount of significant digits until tests can pass across different compilers.

@fwitte, good catch. Just after replying, I realized the silly mistake and fixed it up.

Thank you everyone for the great effort in the review process! It has really helped make the software more robust and the paper a better read.
