### 2020-11-24 BioSTEAMDevelopmentGroup/thermosteam#21 @fwitte
Closing this then, thanks for your explanation.
