### 2020-11-24 9f2dd2f
remove rtruediv; which doesnt make sense
