### 2020-11-24 #2814 @kyleniemeyer
Hi @yoelcortes, I just have few edits needed for the paper:

- Please add an explicit Statement of Need section, which is a [required element of JOSS papers](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) (sorry for not catching that before, usually whedon automatically calls that out upon submission). You might be able to just move some existing text for this.
- please add commas after `e.g.` throughout, as in `e.g., `. Also, you should remove `etc.` in the parenthetical in paragraph three, since e.g. and etc. are redundant
- the RefProp citation has an extraneous `https://doi.org/` in the DOI field
- for the Chemicals and CoolProp software citations, think you should separate the author names so that the reference list handles alphabetical order properly. For example, for Chemicals, the BibTeX entry author field should be `author = {Caleb Bell and {contributors}}`, and similar for CoolProp. This will treat the first author name correctly, and then "contributors" will be kept as a single, separate word.
- For CoolProp, I know there is a paper that should also be cited; here is the BibTeX entry:

```
@article{coolprop,
    author = {Bell, Ian H. and Wronski, Jorrit and Quoilin, Sylvain and Lemort, Vincent},
    title = {Pure and Pseudo-pure Fluid Thermophysical Property Evaluation and
             the Open-Source Thermophysical Property Library CoolProp},
    journal = {Industrial \& Engineering Chemistry Research},
    volume = {53},
    number = {6},
    pages = {2498--2508},
    year = {2014},
    doi = {10.1021/ie4033999},
    }
```
