### 2020-11-25 dc02039
edits to include statement of need, and corrections to bibbliography
