+++ docs/paper/paper.md
@@ -24,48 +24,74 @@ bibliography: paper.bib
 
 # Summary
 
-`Thermosteam` is a standalone thermodynamic engine capable of solving mass and 
+`Thermosteam` is a thermodynamic engine capable of solving mass and 
 energy balances, estimating mixture properties, solving thermodynamic phase 
-equilibria, and modeling stoichiometric reactions. Before getting started with 
-thermodynamic calculations, `Thermosteam` first creates a thermodynamic property 
-package that defines the models used to estimate mixture properties and phase 
-equilibria from both user-defined chemicals and databanks from the `chemicals` 
-library [@chemicals], an open-source compilation of data and functions for the 
-estimation of pure component chemical and mixture properties. Roughly 20,000 chemicals 
-with temperature- and pressure-dependent property data are included in the 
-`chemicals` library. `Thermosteam` builds upon `chemicals` with a robust and 
-flexible framework that facilitates the creation of property packages. 
-Users may easily define new models and pseudo-chemicals that are compatible with
-all of `Thermosteam`'s features. 
+equilibria, and modeling stoichiometric reactions. All chemical data in 
+`Thermosteam` is imported from the `chemicals` library [@chemicals], an 
+open-source compilation of data and functions for the estimation of pure 
+component chemical and mixture properties. `Thermosteam`'s fast and flexible
+platform has enabled the evaluation of conceptual and emerging biochemical production
+processes. The Biorefinery Simulation and Techno-Economic Analysis Modules 
+(BioSTEAM) — capable of modeling reactors, distillation columns, heat exchangers, 
+and other unit operations — has adopted `Thermosteam` as its premier 
+thermodynamic engine [@BioSTEAM]. Published biorefinery designs modeled in 
+BioSTEAM implement thermodynamic property packages created with `Thermosteam` 
+[@Bioindustrial-Park], including a cornstover biorefinery for the production of 
+cellulosic ethanol, a lipid-cane biorefinery for the co-production of ethanol 
+and biodiesel, and a wheatstraw biorefinery for the production of cellulosic 
+ethanol [@BioSTEAM;@Sanchis].
 
-The overarching goal of `Thermosteam` is to enable the rigorous design and 
+# Statement of need
+
+The overarching goal of `Thermosteam` is to aid the rigorous design and 
 simulation of chemical production processes, whereby low value feedstocks are
 converted to high value products via chemical reactions and thermodynamic-driven
 separations. For example, modeling the separation of volatile chemicals from 
-heavier ones in a distillation column (e.g. distilling ethanol from water),
+heavier ones in a distillation column (e.g., distilling ethanol from water),
 requires vapor-liquid phase equilibrium calculations to predict how well volatile 
 chemicals selectively partition into the vapor phase. Additionally, fluid 
 viscosities, densities, and surface tensions are required to appropriately design 
 a distillation column that can achieve a specified recovery of chemicals 
-[@Perry]. The Biorefinery Simulation and Techno-Economic Analysis 
-Modules (BioSTEAM) — capable of modeling distillation columns, reactors,
-heat exchangers, and other unit operations — has adopted `Thermosteam` 
-as its premier thermodynamic engine [@BioSTEAM]. Published designs modeled in 
-BioSTEAM implement property packages created with `Thermosteam` 
-[@Bioindustrial-Park], including a cornstover biorefinery for the production of 
-cellulosic ethanol, a lipid-cane biorefinery for the co-production of ethanol 
-and biodiesel, a sugarcane biorefinery for the production of bioethanol, and a
-wheatstraw biorefinery for the production of cellulosic ethanol [@BioSTEAM;@Sanchis].
+[@Perry]. 
+
+Several open-source libraries in Python have comparable capabilities to 
[…]
