+++ docs/paper/paper.bib
@@ -37,7 +37,7 @@
 }
 
 @misc{chemicals,
-  author = {{​Caleb Bell and Contributors}},
+  author = {Caleb Bell and {contributors}},
   title = {chemicals: Chemical properties component of Chemical Engineering Design Library (ChEDL)},
   year = {2020},
   publisher = {​GitHub},
@@ -45,21 +45,24 @@
   url = {https://github.com/CalebBell/chemicals}
 }
 
-@misc{CoolProp,
-  author = {{​Ian Bell and Contributors}},
-  title = {CoolProp: Thermophysical properties for the masses},
-  year = {2020},
-  publisher = {​GitHub},
-  journal = {​GitHub repository},
-  url = {https://github.com/CoolProp/CoolProp}
-}
+@article{coolprop,
+    author = {Bell, Ian H. and Wronski, Jorrit and Quoilin, Sylvain and Lemort, Vincent},
+    title = {Pure and Pseudo-pure Fluid Thermophysical Property Evaluation and
+             the Open-Source Thermophysical Property Library CoolProp},
+    journal = {Industrial \& Engineering Chemistry Research},
+    volume = {53},
+    number = {6},
+    pages = {2498--2508},
+    year = {2014},
+    doi = {10.1021/ie4033999},
+    }
 
 @misc{REFPROP,
-  Title  = {{NIST Standard Reference Database 23: Reference Fluid Thermodynamic and Transport Properties-REFPROP, Version 10.0, National Institute of Standards and Technology}},
-  Author = {E. W. Lemmon and I.H. Bell and M. L. Huber and M. O. McLinden},
-  Year   = {2018},
-  Doi    = {https://doi.org/10.18434/T4/1502528},
-  Url    = {https://www.nist.gov/srd/refprop}
+  title  = {{NIST Standard Reference Database 23: Reference Fluid Thermodynamic and Transport Properties-REFPROP, Version 10.0, National Institute of Standards and Technology}},
+  author = {E. W. Lemmon and I.H. Bell and M. L. Huber and M. O. McLinden},
+  year   = {2018},
+  doi    = {10.18434/T4/1502528},
+  url    = {https://www.nist.gov/srd/refprop}
 }
 
 @software{Cantera,
