+++ docs/paper/paper.bib
@@ -45,7 +45,7 @@
   url = {https://github.com/CalebBell/chemicals}
 }
 
-@article{coolprop,
+@article{CoolProp,
     author = {Bell, Ian H. and Wronski, Jorrit and Quoilin, Sylvain and Lemort, Vincent},
     title = {Pure and Pseudo-pure Fluid Thermophysical Property Evaluation and
              the Open-Source Thermophysical Property Library CoolProp},
