### 2020-11-25 #2814 @yoelcortes
Hi @kyleniemeyer,

Thanks for the coolprop reference. I added the statement of need along with another section called "roadmap". I mainly shuffled sentences around, but I also added/removed a few here and there.

Please let me know if you have any more comments,
Thanks!
