+++ thermosteam/equilibrium/dew_point.py
@@ -68,7 +68,7 @@ class DewPoint:
     __slots__ = ('chemicals', 'phi', 'gamma', 'IDs',
                  'pcf', 'Psats', 'P', 'T', 'x',
                  'Tmin', 'Tmax', 'Pmin', 'Pmax')
-    Tmin_default = 200.
+    Tmin_default = 150.
     _cached = {}
     def __init__(self, chemicals=(), thermo=None):
         thermo = settings.get_default_thermo(thermo)
